Separate node state composition and lifecycle owners
Keep NodeState as the handle facade; isolate storage, restore, pruning,
checkpoint, event publication, and index startup/shutdown. Group unit
contracts with shared fixtures, remove one duplicate empty-peer test,
and extend the concrete-backend ownership regression to all new owners.

Preserve public handle paths, epoch durability, prune watermarks,
shutdown ordering, and authoritative Chainstate ownership.
