+++ crates/node/src/state.rs
@@ -1,52 +1,51 @@
-//! Shared node state aggregating subsystem handles.
+//! Shared node runtime state and capability handles.
 //!
-//! V1 keeps this deliberately minimal: it owns the resolved [`NodeConfig`], the
-//! data-directory path, and the open chainstate storage backend. Subsystem
-//! wiring (chain / utxo / mempool
-//! / index / p2p / rpc / script_index) parks here as the integration point matures.
+//! Construction, recovery, persistence, event publication, pruning, and index
+//! lifecycle are isolated below; this facade retains the public handle API.
 
-use anyhow::{Context as _, Result, bail};
+mod checkpoint;
+mod events;
+mod index;
+mod open;
+mod prune;
+mod restore;
+mod storage;
 
+#[cfg(test)]
+pub(crate) use restore::ResumeSource;
+
+use crate::ApplyError;
+use crate::NodeConfig;
 use arc_swap::ArcSwapOption;
-
-use bitcoin_rs_chain::{BlockBodyMetadata, BlockBodySource, TipSnapshot};
-
-use bitcoin_rs_mempool::{Mempool, MempoolLimits};
-
-use bitcoin_rs_primitives::{
-    Block, Hash256, Tx, Txid, chain_constants::CORE_REORG_SAFETY_MARGIN, deserialize,
-};
-
-use bitcoin_rs_rpc::context::{
-    BlockLog, NetworkState, PruneResult, PruneService, PruneServiceError, PruneStatus,
-};
-
-use bitcoin_rs_storage::{
-    ColumnFamily, FlatFileBlockStore, KvStore, StorageBackend, WriteBatch,
-    pruning::{PrunePolicy, reclaim_staged_flat_block_files, stage_block_and_undo_prune},
-};
-
+use bitcoin_rs_chain::BlockBodySource;
+use bitcoin_rs_chain::TipSnapshot;
+use bitcoin_rs_mempool::Mempool;
+use bitcoin_rs_primitives::Block;
+use bitcoin_rs_primitives::Tx;
+use bitcoin_rs_primitives::Txid;
+use bitcoin_rs_rpc::context::BlockLog;
+use bitcoin_rs_rpc::context::NetworkState;
+use bitcoin_rs_rpc::context::PruneService;
 use bitcoin_rs_utxo::UtxoSet;
-
-use core::mem::size_of;
-
-use crate::{ApplyError, NodeConfig};
-
-use crossbeam_channel::{Receiver, Sender};
-
+use crossbeam_channel::Receiver;
+use crossbeam_channel::Sender;
+pub use events::ChainEventHint;
+pub use events::ChainEventPublisher;
+pub use events::ChainSnapshot;
+pub use events::HintKind;
 use hashbrown::HashMap;
-
-use parking_lot::{Mutex, RwLock};
-
-use std::{
-    io::{self, Write as _},
-    path::{Path, PathBuf},
-    sync::{
-        Arc,
-        atomic::{AtomicBool, AtomicU32, AtomicU64, Ordering},
-    },
-    time::Duration,
-};
+use index::TxIndexSpawn;
+use parking_lot::Mutex;
+use parking_lot::RwLock;
+pub use prune::NodePruneService;
+use std::path::Path;
+use std::path::PathBuf;
+use std::sync::Arc;
+use std::sync::atomic::AtomicBool;
+use std::sync::atomic::AtomicU32;
+use std::sync::atomic::AtomicU64;
+use storage::NodeStorage;
+use storage::StoredBlockBodySource;
 
 // One active generation of outbound requests is enough to keep the drain fed;
 // extra backlog is overload and must fail fast at producers.
@@ -62,6 +61,7 @@ pub(crate) const P2P_OUTBOUND_QUEUE_LIMIT: usize = 8;
 // in-flight request window (`PENDING_BUDGET` = 256) so honest delivery, which
 // wakes the drain on every block, is never throttled.
 pub(crate) const INBOUND_BLOCK_CHANNEL_LIMIT: usize = 512;
+
 // Bounds chain-event hints between the block-apply commit path and
 // reconciliation consumers (#77). Hints are wake-ups, never data: a consumer
 // that misses one recovers by reconciling `ChainSnapshot` against its own
@@ -69,1068 +69,13 @@ pub(crate) const INBOUND_BLOCK_CHANNEL_LIMIT: usize = 512;
 // inbound-block bound so both channels share the same flood posture; a full
 // channel drops the hint and never blocks the commit path.
 pub(crate) const CHAIN_HINT_CHANNEL_LIMIT: usize = INBOUND_BLOCK_CHANNEL_LIMIT;
+
 // Bounds inbound peer transactions between the per-peer listener threads and
 // the single ingress consumer. A full channel applies TCP backpressure to
 // that peer's read loop; other peers keep their own threads. Sized to absorb
 // a burst of honest `tx` deliveries without stalling header/block traffic on
 // the same connection under normal load.
 pub(crate) const INBOUND_TX_CHANNEL_LIMIT: usize = 1_024;
-
-/// A coherent, non-torn view of the applied chain tip.
-///
-/// The only writer replaces the whole cell under one `RwLock`, so a reader
-/// never observes a torn mix of two commit points. This is a live value: it is
-/// never persisted per-event. `epoch` changes only across process restarts,
-/// `sequence` advances once per committed connect/disconnect (`0` means no
-/// committed event yet this run), and the tip fields name the block that
-/// sequence was advanced for.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct ChainSnapshot {
-    /// Persisted process epoch, strictly monotonic per data dir.
-    pub epoch: u64,
-    /// Commit counter; starts at `1` on the first `record` of a run.
-    pub sequence: u64,
-    /// Applied tip block hash (genesis hash before the first commit).
-    pub tip_hash: Hash256,
-    /// Applied tip height (`0` at genesis).
-    pub tip_height: u32,
-}
-
-/// Which committed chain event a [`ChainEventHint`] describes.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum HintKind {
-    /// A block was committed onto the tip.
-    Connected,
-    /// The tip moved back to its parent during a disconnect/reorg.
-    Disconnected,
-}
-
-/// Wake-up for reconciliation consumers: one committed connect or disconnect.
-///
-/// Hints are not a replay log and carry no payload to apply. A dropped hint is
-/// not a bug — it loses only the wake-up, and the consumer recovers by
-/// reconciling a fresh [`ChainSnapshot`] against its own cursor using the
-/// chain itself: ancestry via `BlockTree::active_node_at_height` and
-/// `BlockTree::find_common_ancestor` (crates/chain), bodies via
-/// `BlockBodyStore::load_block_body` (`crate::apply`). The `epoch` field is
-/// what makes a persisted consumer cursor `(epoch, sequence)` stale on
-/// restart.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct ChainEventHint {
-    /// Whether the block was added to or removed from the tip.
-    pub kind: HintKind,
-    /// Height of the block the event committed.
-    pub height: u32,
-    /// Hash of the block the event committed.
-    pub hash: Hash256,
-    /// Process epoch the event belongs to.
-    pub epoch: u64,
-    /// Commit-counter value assigned to this event.
-    pub sequence: u64,
-}
-
-/// Single write path for chain events: [`Self::record`] advances the commit
-/// sequence, replaces the snapshot cell, then emits the hint, in that order.
-///
-/// A consumer woken by a hint therefore always reads a snapshot at least as
-/// fresh as the hint. Production wiring goes through `NodeState::open`;
-/// [`Self::detached`] exists for `Chainstate` composition in tests.
-pub struct ChainEventPublisher {
-    epoch: u64,
-    sequence: AtomicU64,
-    snapshot: RwLock<ChainSnapshot>,
-    hints: Sender<ChainEventHint>,
-}
-
-impl ChainEventPublisher {
-    fn new(epoch: u64, initial: ChainSnapshot) -> (Self, Receiver<ChainEventHint>) {
-        let (hints, receiver) = crossbeam_channel::bounded(CHAIN_HINT_CHANNEL_LIMIT);
-        (
-            Self {
-                epoch,
-                sequence: AtomicU64::new(0),
-                snapshot: RwLock::new(initial),
-                hints,
-            },
-            receiver,
-        )
-    }
-
-    /// Publisher detached from any node, for test handle composition only.
-    /// Anchors at an empty tip; records still sequence and publish normally.
-    #[must_use]
-    pub fn detached(epoch: u64) -> (Self, Receiver<ChainEventHint>) {
-        Self::new(
-            epoch,
-            ChainSnapshot {
-                epoch,
-                sequence: 0,
-                tip_hash: Hash256::from_le_bytes(&[0; 32]),
-                tip_height: 0,
-            },
-        )
-    }
-
-    /// Returns the process epoch this publisher stamps events with.
-    #[must_use]
-    pub const fn epoch(&self) -> u64 {
-        self.epoch
-    }
-
-    /// Returns the current snapshot without changing any state.
-    #[must_use]
-    pub fn snapshot(&self) -> ChainSnapshot {
-        *self.snapshot.read()
-    }
-
-    /// Records one committed connect or disconnect.
-    ///
-    /// Publication order is fixed: advance the sequence, replace the snapshot
-    /// cell, then `try_send` the hint. A full hint channel drops the hint and
-    /// never blocks or fails the commit path — consumers reconcile from the
-    /// chain, so only the wake-up is lost. Sequence values start at `1`; a
-    /// snapshot with sequence `0` means no committed event yet.
-    pub fn record(&self, kind: HintKind, height: u32, hash: Hash256) -> ChainEventHint {
-        let sequence = self.sequence.fetch_add(1, Ordering::SeqCst) + 1;
-        *self.snapshot.write() = ChainSnapshot {
-            epoch: self.epoch,
-            sequence,
-            tip_hash: hash,
-            tip_height: height,
-        };
-        let hint = ChainEventHint {
-            kind,
-            height,
-            hash,
-            epoch: self.epoch,
-            sequence,
-        };
-        let _ = self.hints.try_send(hint);
-        hint
-    }
-}
-
-const PROCESS_EPOCH_FILE: &str = "process-epoch";
-const PROCESS_EPOCH_LOCK_FILE: &str = ".process-epoch.lock";
-const PROCESS_EPOCH_TEMP: &str = ".process-epoch.tmp";
-// A u64 in decimal is at most 20 digits; the trailing newline makes 21.
-const PROCESS_EPOCH_MAX_BYTES: u64 = 32;
-
-/// Reads the persisted process epoch; `0` when the data dir has none yet.
-///
-/// A corrupt file is an error, not a reset: silently restarting the counter
-/// would let a new run reuse an epoch old consumer cursors live in.
-fn load_process_epoch(dir: &cap_std::fs::Dir) -> Result<u64> {
-    let bytes =
-        match crate::checkpoint_fs::read_file(dir, PROCESS_EPOCH_FILE, PROCESS_EPOCH_MAX_BYTES) {
-            Ok(bytes) => bytes,
-            Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok(0),
-            Err(error) => {
-                return Err(error).with_context(|| format!("read {PROCESS_EPOCH_FILE}"));
-            }
-        };
-    let text = core::str::from_utf8(&bytes)
-        .with_context(|| format!("{PROCESS_EPOCH_FILE} is not valid UTF-8"))?;
-    text.trim().parse::<u64>().with_context(|| {
-        format!("{PROCESS_EPOCH_FILE} is corrupt; refusing to start rather than reuse epochs")
-    })
-}
-
-/// Allocates the next process epoch, durably, before first use.
-///
-/// The persistent lock serializes the complete load → increment → temporary
-/// file sync → rename → data-directory sync transaction across processes.
-/// Keeping its descriptor alive through the final directory sync matters:
-/// opening the data directory does not freeze its namespace or mount topology.
-/// The epoch itself lives outside the re-writable checkpoint tree, so a
-/// checkpoint wipe or resync can never regress it. A crash before the rename
-/// may leave a temporary file; gaps are fine, but reuse is not.
-fn allocate_process_epoch(dir: &cap_std::fs::Dir) -> Result<u64> {
-    use cap_fs_ext::FollowSymlinks;
-    use cap_fs_ext::OpenOptionsFollowExt as _;
-    use cap_fs_ext::OpenOptionsSyncExt as _;
-
-    let mut lock_options = cap_std::fs::OpenOptions::new();
-    lock_options
-        .read(true)
-        .write(true)
-        .create(true)
-        .follow(FollowSymlinks::No)
-        .nonblock(true);
-    let lock = dir
-        .open_with(PROCESS_EPOCH_LOCK_FILE, &lock_options)
-        .with_context(|| format!("open process epoch lock {PROCESS_EPOCH_LOCK_FILE}"))?;
-    let lock_metadata = lock
-        .metadata()
-        .with_context(|| format!("inspect process epoch lock {PROCESS_EPOCH_LOCK_FILE}"))?;
-    if !lock_metadata.is_file() {
-        bail!("process epoch lock {PROCESS_EPOCH_LOCK_FILE} is not a regular file");
-    }
-    rustix::fs::flock(&lock, rustix::fs::FlockOperation::LockExclusive)
-        .with_context(|| format!("lock process epoch file {PROCESS_EPOCH_LOCK_FILE}"))?;
-
-    let epoch = load_process_epoch(dir)?
-        .checked_add(1)
-        .context("process epoch counter exhausted")?;
-    let bytes = format!("{epoch}\n").into_bytes();
-    match dir.remove_file(PROCESS_EPOCH_TEMP) {
-        Ok(()) => {}
-        Err(error) if error.kind() == io::ErrorKind::NotFound => {}
-        Err(error) => {
-            return Err(error).with_context(|| format!("remove stale {PROCESS_EPOCH_TEMP}"));
-        }
-    }
-
-    let allocation = (|| -> Result<()> {
-        let mut file = crate::checkpoint_fs::create_file(dir, PROCESS_EPOCH_TEMP)
-            .with_context(|| format!("create {PROCESS_EPOCH_TEMP}"))?;
-        file.write_all(&bytes)
-            .with_context(|| format!("write {PROCESS_EPOCH_TEMP}"))?;
-        file.sync_all()
-            .with_context(|| format!("sync {PROCESS_EPOCH_TEMP}"))?;
-        drop(file);
-        dir.rename(PROCESS_EPOCH_TEMP, dir, PROCESS_EPOCH_FILE)
-            .with_context(|| format!("publish {PROCESS_EPOCH_FILE}"))?;
-        crate::checkpoint_fs::sync_dir(dir)
-            .context("sync data dir after allocating the process epoch")
-    })();
-    if allocation.is_err() {
-        let _ = dir.remove_file(PROCESS_EPOCH_TEMP);
-    }
-    allocation?;
-
-    // `lock` intentionally remains live until after the directory durability
-    // barrier above. Dropping it here releases the cross-process transaction.
-    drop(lock);
-    Ok(epoch)
-}
-
-struct NodeStorage {
-    backend: StorageBackend,
-    undo_store: Arc<dyn crate::apply::UndoStore>,
-    block_body_store: Arc<dyn bitcoin_rs_storage::block_body::BlockBodyStore>,
-    deferred: Arc<dyn DeferredChainstateServices>,
-    #[cfg(test)]
-    test_store: Arc<dyn TestStoreAccess>,
-}
-
-struct ChainstateComposer {
-    backend: StorageBackend,
-    block_files: Arc<FlatFileBlockStore>,
-}
-
-impl crate::storage_backend::StoreConsumer for ChainstateComposer {
-    type Output = NodeStorage;
-    type Error = bitcoin_rs_storage::StorageError;
-
-    fn consume<S>(
-        self,
-        store: Arc<S>,
-    ) -> core::result::Result<Self::Output, bitcoin_rs_storage::StorageError>
-    where
-        S: KvStore,
-    {
-        let deferred: Arc<dyn DeferredChainstateServices> = Arc::new(ChainstateStoreServices {
-            store: Arc::clone(&store),
-        });
-        Ok(NodeStorage {
-            backend: self.backend,
-            undo_store: Arc::new(crate::apply::KvUndoStore::new(Arc::clone(&store))),
-            block_body_store: Arc::new(bitcoin_rs_storage::block_body::IndexedBlockBodyStore::new(
-                Arc::clone(&store),
-                self.block_files,
-            )),
-            deferred,
-            #[cfg(test)]
-            test_store: Arc::new(TestStore { store }),
-        })
-    }
-}
-
-impl NodeStorage {
-    /// Opens the configured backend for the chainstate namespace with its
-    /// cache share from the process budget.
-    fn open(
-        config: &NodeConfig,
-        chainstate_cache_bytes: u64,
-        block_files: Arc<FlatFileBlockStore>,
-    ) -> Result<Self> {
-        let chainstate_dir = config.data_dir.join("chainstate");
-        std::fs::create_dir_all(&chainstate_dir)
-            .with_context(|| format!("create chainstate_dir {}", chainstate_dir.display()))?;
-
-        let backend = config.storage.backend;
-        crate::storage_backend::open_chainstate(
-            backend,
-            &chainstate_dir,
-            Some(chainstate_cache_bytes),
-            ChainstateComposer {
-                backend,
-                block_files,
-            },
-        )
-        .map_err(anyhow::Error::new)
-    }
-
-    const fn kind(&self) -> &'static str {
-        self.backend.as_str()
-    }
-
-    fn prune_service(
-        &self,
-        block_files: &Arc<FlatFileBlockStore>,
-        block_body_store: &Arc<dyn bitcoin_rs_storage::block_body::BlockBodyStore>,
-        blocks: Arc<RwLock<BlockLog>>,
-        transactions: Arc<RwLock<HashMap<Txid, Tx>>>,
-        authority: crate::apply::PruneAuthority,
-        durable_tip_height: &Arc<AtomicU32>,
-    ) -> Result<Arc<dyn PruneService>> {
-        self.deferred.prune_service(
-            Arc::clone(block_files),
-            Arc::clone(block_body_store),
-            blocks,
-            transactions,
-            authority,
-            Arc::clone(durable_tip_height),
-        )
-    }
-
-    fn block_body_store(&self) -> Arc<dyn bitcoin_rs_storage::block_body::BlockBodyStore> {
-        Arc::clone(&self.block_body_store)
-    }
-
-    /// Builds the undo store for the configured backend.
-    ///
-    /// Mandatory rather than optional: without undo records the node cannot
-    /// disconnect a block, so it could advance its tip into a chain it is
-    /// unable to leave.
-    fn undo_store(&self) -> Arc<dyn crate::apply::UndoStore> {
-        Arc::clone(&self.undo_store)
-    }
-
-    fn journal_writer(
-        &self,
-        dir: cap_std::fs::Dir,
-        bootstrap: JournalBootstrap,
-    ) -> Result<crate::chainstate_journal::SharedJournalWriter> {
-        self.deferred.journal_writer(dir, bootstrap)
-    }
-
-    #[cfg(test)]
-    fn stored_prune_body(
-        &self,
-        height: u32,
-        hash: bitcoin_rs_primitives::Hash256,
-    ) -> Result<Option<Vec<u8>>> {
-        let key = bitcoin_rs_storage::pruning::block_body_key(height, hash);
-        Ok(self
-            .test_store
-            .get(bitcoin_rs_storage::pruning::BLOCK_DATA_CF, &key)?)
-    }
-
-    #[cfg(test)]
-    fn stored_prune_undo(
-        &self,
-        height: u32,
-        hash: bitcoin_rs_primitives::Hash256,
-    ) -> Result<Option<Vec<u8>>> {
-        let key = bitcoin_rs_storage::pruning::block_undo_key(height, hash);
-        Ok(self.test_store.get(ColumnFamily::UndoData, &key)?)
-    }
-
-    #[cfg(test)]
-    fn write_test_rows(&self, rows: &[(ColumnFamily, Vec<u8>, Vec<u8>)]) -> Result<()> {
-        self.test_store.write_rows(rows).map_err(anyhow::Error::new)
-    }
-
-    #[cfg(test)]
-    fn read_test_row(&self, cf: ColumnFamily, key: &[u8]) -> Result<Option<Vec<u8>>> {
-        self.test_store.get(cf, key).map_err(anyhow::Error::new)
-    }
-}
-
-#[derive(Clone, Copy)]
-struct JournalBootstrap {
-    open_existing: bool,
-    base_generation: u64,
-    height: u32,
-    block_hash: [u8; 32],
-    prev_hash: [u8; 32],
-    chain_tx_count: u64,
-    config: crate::config::ChainstateJournalConfig,
-}
-
-/// Capabilities whose inputs become available after the chainstate store is
-/// opened. This is a composition seam, not a storage API: concrete reads,
-/// batches, and durability remain generic over `KvStore` below it.
-trait DeferredChainstateServices: Send + Sync {
-    fn prune_service(
-        &self,
-        block_files: Arc<FlatFileBlockStore>,
-        block_body_store: Arc<dyn bitcoin_rs_storage::block_body::BlockBodyStore>,
-        blocks: Arc<RwLock<BlockLog>>,
-        transactions: Arc<RwLock<HashMap<Txid, Tx>>>,
-        authority: crate::apply::PruneAuthority,
-        durable_tip_height: Arc<AtomicU32>,
-    ) -> Result<Arc<dyn PruneService>>;
-
-    fn journal_writer(
-        &self,
-        dir: cap_std::fs::Dir,
-        bootstrap: JournalBootstrap,
-    ) -> Result<crate::chainstate_journal::SharedJournalWriter>;
-}
-
-struct ChainstateStoreServices<S> {
-    store: Arc<S>,
-}
-
-impl<S: KvStore> DeferredChainstateServices for ChainstateStoreServices<S> {
-    fn prune_service(
-        &self,
-        block_files: Arc<FlatFileBlockStore>,
-        block_body_store: Arc<dyn bitcoin_rs_storage::block_body::BlockBodyStore>,
-        blocks: Arc<RwLock<BlockLog>>,
-        transactions: Arc<RwLock<HashMap<Txid, Tx>>>,
-        authority: crate::apply::PruneAuthority,
-        durable_tip_height: Arc<AtomicU32>,
-    ) -> Result<Arc<dyn PruneService>> {
-        Ok(Arc::new(NodePruneService::new(
-            Arc::clone(&self.store),
-            block_files,
-            block_body_store,
-            blocks,
-            transactions,
-            authority,
-            durable_tip_height,
-        )?))
-    }
-
-    fn journal_writer(
-        &self,
-        dir: cap_std::fs::Dir,
-        bootstrap: JournalBootstrap,
-    ) -> Result<crate::chainstate_journal::SharedJournalWriter> {
-        build_journal_writer(dir, Arc::clone(&self.store), bootstrap)
-    }
-}
-
-#[cfg(test)]
-trait TestStoreAccess: Send + Sync {
-    fn get(
-        &self,
-        cf: ColumnFamily,
-        key: &[u8],
-    ) -> core::result::Result<Option<Vec<u8>>, bitcoin_rs_storage::StorageError>;
-
-    fn write_rows(
-        &self,
-        rows: &[(ColumnFamily, Vec<u8>, Vec<u8>)],
-    ) -> core::result::Result<(), bitcoin_rs_storage::StorageError>;
-}
-
-#[cfg(test)]
-struct TestStore<S> {
-    store: Arc<S>,
-}
-
-#[cfg(test)]
-impl<S: KvStore> TestStoreAccess for TestStore<S> {
-    fn get(
-        &self,
-        cf: ColumnFamily,
-        key: &[u8],
-    ) -> core::result::Result<Option<Vec<u8>>, bitcoin_rs_storage::StorageError> {
-        self.store.get(cf, key)
-    }
-
-    fn write_rows(
-        &self,
-        rows: &[(ColumnFamily, Vec<u8>, Vec<u8>)],
-    ) -> core::result::Result<(), bitcoin_rs_storage::StorageError> {
-        let mut batch = self.store.new_batch();
-        for (cf, key, value) in rows {
-            batch.put(*cf, key, value);
-        }
-        self.store.write(batch)
-    }
-}
-
-fn build_journal_writer<S: KvStore + 'static>(
-    dir: cap_std::fs::Dir,
-    store: Arc<S>,
-    bootstrap: JournalBootstrap,
-) -> Result<crate::chainstate_journal::SharedJournalWriter> {
-    let mut writer = if bootstrap.open_existing {
-        crate::chainstate_journal::JournalWriter::open(dir, store)?
-    } else {
-        crate::chainstate_journal::JournalWriter::initialize(
-            dir,
-            store,
-            bootstrap.base_generation,
-            (0, 0),
-            bootstrap.height,
-            bootstrap.block_hash,
-            bootstrap.prev_hash,
-            bootstrap.chain_tx_count,
-        )?
-    };
-    writer.configure(
-        bootstrap.config.blocks,
-        Duration::from_secs(bootstrap.config.seconds),
-        bootstrap.config.rotate_mib,
-        bootstrap.config.max_journal_mib,
-        bootstrap.config.max_lag_blocks,
-        Duration::from_secs(bootstrap.config.max_lag_seconds),
-    )?;
-    Ok(crate::chainstate_journal::shared_journal_writer(writer))
-}
-
-struct StoredBlockBodySource {
-    store: Arc<dyn bitcoin_rs_storage::block_body::BlockBodyStore>,
-}
-
-impl StoredBlockBodySource {
-    fn new(store: Arc<dyn bitcoin_rs_storage::block_body::BlockBodyStore>) -> Self {
-        Self { store }
-    }
-}
-
-impl BlockBodySource for StoredBlockBodySource {
-    fn block_body(&self, height: u32, hash: bitcoin_rs_primitives::BlockHash) -> Option<Vec<u8>> {
-        self.store.load_block_body(height, hash.0).ok().flatten()
-    }
-
-    fn disk_usage(&self) -> Option<u64> {
-        self.store.disk_usage()
-    }
-
-    fn block_body_range(
-        &self,
-        height: u32,
-        hash: bitcoin_rs_primitives::BlockHash,
-        offset: u32,
-        len: u32,
-    ) -> Option<Vec<u8>> {
-        // `None` is overloaded here: it means both "this store cannot slice"
-        // and "the read failed". Callers must treat either as a reason to fall
-        // back to the whole body, so the return type stays — but this is the
-        // hot path for every ScriptIndex history call now, and an I/O error that
-        // silently degrades into a full block scan is exactly the failure that
-        // would otherwise show up only as unexplained latency.
-        match self
-            .store
-            .load_block_body_range(height, hash.0, offset, len)
-        {
-            Ok(bytes) => bytes,
-            Err(error) => {
-                tracing::debug!(
-                    %error,
-                    height,
-                    offset,
-                    len,
-                    "ranged block body read failed; falling back to the whole body"
-                );
-                None
-            }
-        }
-    }
-
-    fn block_body_metadata(
-        &self,
-        height: u32,
-        hash: bitcoin_rs_primitives::BlockHash,
-    ) -> Option<BlockBodyMetadata> {
-        self.store
-            .block_body_metadata(height, hash.0)
-            .ok()
-            .flatten()
-
-    }
-}
-
-const PRUNEHEIGHT_METADATA_KEY: &[u8] = b"node:pruneheight";
-
-/// A checkpoint restore more than this many blocks behind the durable
-/// applied-tip witness is a catastrophic rollback, not a routine resume.
-/// The restore is still accepted — the chainstate is valid — but the node
-/// logs at ERROR and the warning snapshot carries the gap so operators
-/// and RPC consumers can see the node is starting far behind where it was.
-const STALE_RESTORE_ERROR_THRESHOLD: u32 = 1000;
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(crate) enum ResumeSource {
-    Cold,
-    Checkpoint,
-    Journal,
-}
-
-pub(crate) const CHAINSTATE_JOURNAL_DIR: &str = crate::chainstate_journal::JOURNAL_DIR_NAME;
-
-fn requires_full_revalidation(data_dir: &Path) -> bool {
-    data_dir
-        .join(CHAINSTATE_JOURNAL_DIR)
-        .join(crate::chainstate_journal::FULL_REVALIDATION_MARKER)
-        .is_file()
-}
-
-struct InitialChainstate {
-    utxo: UtxoSet,
-    coin_stats: bitcoin_rs_utxo::stats::CoinStats,
-    tree: bitcoin_rs_chain::BlockTree,
-    applied_tip: Option<TipSnapshot>,
-    chain_tx_count: u64,
-    resume_source: ResumeSource,
-    journal_bootstrap: Option<JournalBootstrap>,
-}
-
-fn reset_journal_dir(data_dir: &Path) -> Result<cap_std::fs::Dir> {
-    let path = data_dir.join(CHAINSTATE_JOURNAL_DIR);
-    match std::fs::remove_dir_all(&path) {
-        Ok(()) => {}
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => {}
-        Err(error) => return Err(error).with_context(|| format!("remove {}", path.display())),
-    }
-    std::fs::create_dir_all(&path).with_context(|| format!("create {}", path.display()))?;
-    crate::checkpoint_fs::open_data_dir(&path).with_context(|| format!("open {}", path.display()))
-}
-
-fn open_journal_dir(data_dir: &Path) -> Result<cap_std::fs::Dir> {
-    let path = data_dir.join(CHAINSTATE_JOURNAL_DIR);
-    std::fs::create_dir_all(&path).with_context(|| format!("create {}", path.display()))?;
-    crate::checkpoint_fs::open_data_dir(&path).with_context(|| format!("open {}", path.display()))
-}
-
-fn checkpoint_bootstrap(
-    restored: &crate::checkpoint::RestoredChainstate,
-    config: crate::config::ChainstateJournalConfig,
-    open_existing: bool,
-) -> Result<JournalBootstrap> {
-    let node = restored.tree.node(restored.applied_tip.tip_id)?;
-    let prev_hash = match node.parent {
-        Some(parent) => restored.tree.node(parent)?.hash.to_le_bytes(),
-        None => [0_u8; 32],
-    };
-    Ok(JournalBootstrap {
-        open_existing,
-        base_generation: restored.generation,
-        height: restored.applied_tip.height,
-        block_hash: restored.applied_tip.hash.to_le_bytes(),
-        prev_hash,
-        chain_tx_count: restored.chain_tx_count,
-        config,
-    })
-}
-
-fn restored_initial(
-    restored: crate::checkpoint::RestoredChainstate,
-    config: crate::config::ChainstateJournalConfig,
-    open_existing: bool,
-    resume_source: ResumeSource,
-) -> Result<InitialChainstate> {
-    let journal_bootstrap = config
-        .enabled
-        .then(|| checkpoint_bootstrap(&restored, config, open_existing))
-        .transpose()?;
-    Ok(InitialChainstate {
-        utxo: restored.utxo,
-        coin_stats: restored.coin_stats,
-        tree: restored.tree,
-        applied_tip: Some(restored.applied_tip),
-        chain_tx_count: restored.chain_tx_count,
-        resume_source,
-        journal_bootstrap,
-    })
-}
-
-fn cold_initial_chainstate(
-    config: &NodeConfig,
-    journal_config: crate::config::ChainstateJournalConfig,
-    reset_journal: bool,
-) -> Result<InitialChainstate> {
-    if journal_config.enabled && reset_journal {
-        drop(reset_journal_dir(&config.data_dir)?);
-    }
-    Ok(InitialChainstate {
-        utxo: UtxoSet::new(),
-        coin_stats: bitcoin_rs_utxo::stats::CoinStats::default(),
-        tree: bitcoin_rs_chain::BlockTree::new(),
-        applied_tip: None,
-        chain_tx_count: 0,
-        resume_source: ResumeSource::Cold,
-        journal_bootstrap: journal_config.enabled.then_some(JournalBootstrap {
-            open_existing: false,
-            base_generation: 0,
-            height: 0,
-            block_hash: config.network.genesis_block_hash().to_le_bytes(),
-            prev_hash: [0_u8; 32],
-            chain_tx_count: 1,
-            config: journal_config,
-        }),
-    })
-}
-
-fn prepare_initial_chainstate(
-    checkpoint_load: crate::checkpoint::CheckpointLoad,
-    checkpoint_data_dir: &cap_std::fs::Dir,
-    checkpoint_config: crate::checkpoint::HeaderCheckpointConfig,
-    config: &NodeConfig,
-) -> Result<InitialChainstate> {
-    let journal_config = config.chainstate_journal;
-    if requires_full_revalidation(&config.data_dir) {
-        metrics::counter!(
-            "node.chainstate_journal.fallback_total",
-            "reason" => "full_revalidation_marker"
-        )
-        .increment(1);
-        tracing::warn!(
-            restore_source = "cold",
-            reason = "fork_below_checkpoint_base",
-            "chainstate restore requires full validation"
-        );
-        return cold_initial_chainstate(config, journal_config, false);
-    }
-    let crate::checkpoint::CheckpointLoad::Complete(restored) = checkpoint_load else {
-        if journal_config.enabled {
-            metrics::counter!(
-                "node.chainstate_journal.fallback_total",
-                "reason" => "no_checkpoint"
-            )
-            .increment(1);
-        }
-        tracing::info!(
-            restore_source = "cold",
-            reason = "no_complete_checkpoint",
-            "chainstate restore selected"
-        );
-        return cold_initial_chainstate(config, journal_config, true);
-    };
-    let restored = *restored;
-    if !journal_config.enabled {
-        tracing::info!(
-            restore_source = "checkpoint",
-            height = restored.applied_tip.height,
-            hash = %restored.applied_tip.hash,
-            chain_tx_count = restored.chain_tx_count,
-            reason = "journal_disabled",
-            "chainstate restore selected"
-        );
-        return restored_initial(restored, journal_config, false, ResumeSource::Checkpoint);
-    }
-
-    replay_checkpoint_journal(
-        restored,
-        checkpoint_data_dir,
-        checkpoint_config,
-        config,
-        journal_config,
-    )
-}
-
-fn replay_checkpoint_journal(
-    restored: crate::checkpoint::RestoredChainstate,
-    checkpoint_data_dir: &cap_std::fs::Dir,
-    checkpoint_config: crate::checkpoint::HeaderCheckpointConfig,
-    config: &NodeConfig,
-    journal_config: crate::config::ChainstateJournalConfig,
-) -> Result<InitialChainstate> {
-    let base_generation = restored.generation;
-    let base_height = restored.applied_tip.height;
-    let journal_dir = open_journal_dir(&config.data_dir)?;
-    let replay_started = std::time::Instant::now();
-    let replay = crate::chainstate_journal::replay_from_journal(
-        &journal_dir,
-        base_generation,
-        restored.tree,
-        restored.utxo,
-        restored.coin_stats,
-        restored.applied_tip,
-        restored.chain_tx_count,
-    );
-    let replay_seconds = replay_started.elapsed().as_secs_f64();
-    metrics::histogram!("node.chainstate_journal.replay_seconds").record(replay_seconds);
-    drop(journal_dir);
-    match replay {
-        crate::chainstate_journal::ReplayOutcome::Replayed(replayed) => {
-            let replayed_records = replayed.applied_tip.height.saturating_sub(base_height);
-            let (restore_source, resume_source) = if replayed_records == 0 {
-                ("checkpoint", ResumeSource::Checkpoint)
-            } else {
-                ("journal", ResumeSource::Journal)
-            };
-            tracing::info!(
-                restore_source,
-                checkpoint_generation = base_generation,
-                checkpoint_height = base_height,
-                height = replayed.applied_tip.height,
-                hash = %replayed.applied_tip.hash,
-                replayed_records,
-                chain_tx_count = replayed.chain_tx_count,
-                replay_seconds,
-                "chainstate restore selected"
-            );
-            let bootstrap = JournalBootstrap {
-                open_existing: true,
-                base_generation,
-                height: replayed.applied_tip.height,
-                block_hash: replayed.applied_tip.hash.to_le_bytes(),
-                prev_hash: [0_u8; 32],
-                chain_tx_count: replayed.chain_tx_count,
-                config: journal_config,
-            };
-            Ok(InitialChainstate {
-                utxo: replayed.utxo,
-                coin_stats: replayed.coin_stats,
-                tree: replayed.tree,
-                applied_tip: Some(replayed.applied_tip),
-                chain_tx_count: replayed.chain_tx_count,
-                resume_source,
-                journal_bootstrap: Some(bootstrap),
-            })
-        }
-        crate::chainstate_journal::ReplayOutcome::Fallback(error) => {
-            let reason = error.reason();
-            metrics::counter!(
-                "node.chainstate_journal.fallback_total",
-                "reason" => reason
-            )
-            .increment(1);
-            if error.is_checksum_failure() {
-                metrics::counter!("node.chainstate_journal.checksum_failures_total").increment(1);
-            }
-            tracing::warn!(
-                restore_source = "checkpoint",
-                checkpoint_generation = base_generation,
-                checkpoint_height = base_height,
-                reason,
-                %error,
-                replay_seconds,
-                "chainstate journal rejected; checkpoint recovery selected"
-            );
-            drop(reset_journal_dir(&config.data_dir)?);
-            let reloaded = crate::checkpoint::load_checkpoint_from_dir(
-                checkpoint_data_dir,
-                checkpoint_config,
-            )?;
-            let crate::checkpoint::CheckpointLoad::Complete(reloaded) = reloaded else {
-                bail!("checkpoint disappeared while recovering from journal fallback");
-            };
-            restored_initial(*reloaded, journal_config, false, ResumeSource::Checkpoint)
-        }
-    }
-}
-
-fn load_pruneheight<S: KvStore>(store: &S) -> Result<Option<u32>> {
-    let Some(bytes) = store.get(ColumnFamily::UtxoMeta, PRUNEHEIGHT_METADATA_KEY)? else {
-        return Ok(None);
-    };
-    if bytes.len() != size_of::<u32>() {
-        bail!("invalid persisted pruneheight length {}", bytes.len());
-    }
-    let mut encoded = [0_u8; size_of::<u32>()];
-    encoded.copy_from_slice(&bytes);
-    Ok(Some(u32::from_be_bytes(encoded)))
-}
-
-/// Storage-backed implementation of RPC manual pruning.
-pub struct NodePruneService<S: KvStore> {
-    store: Arc<S>,
-    block_files: Arc<FlatFileBlockStore>,
-    block_body_store: Arc<dyn bitcoin_rs_storage::block_body::BlockBodyStore>,
-    blocks: Arc<RwLock<BlockLog>>,
-    transactions: Arc<RwLock<HashMap<Txid, Tx>>>,
-    authority: crate::apply::PruneAuthority,
-    pruneheight: Mutex<Option<u32>>,
-    /// Height the last clean checkpoint would restore to, 0 when none exists.
-    ///
-    /// Undo pruning is bounded by this, not by the in-memory applied tip, which
-    /// can run far ahead of it.
-    durable_tip_height: Arc<AtomicU32>,
-}
-
-impl<S: KvStore> NodePruneService<S> {
-    /// Creates a manual pruning service over the chainstate store and RPC block cache.
-    pub(crate) fn new(
-        store: Arc<S>,
-        block_files: Arc<FlatFileBlockStore>,
-        block_body_store: Arc<dyn bitcoin_rs_storage::block_body::BlockBodyStore>,
-        blocks: Arc<RwLock<BlockLog>>,
-        transactions: Arc<RwLock<HashMap<Txid, Tx>>>,
-        authority: crate::apply::PruneAuthority,
-        durable_tip_height: Arc<AtomicU32>,
-    ) -> Result<Self> {
-        let pruneheight = load_pruneheight(&*store)?;
-        Ok(Self {
-            store,
-            block_files,
-            block_body_store,
-            blocks,
-            transactions,
-            authority,
-            pruneheight: Mutex::new(pruneheight),
-            durable_tip_height,
-        })
-    }
-}
-
-impl<S: KvStore> PruneService for NodePruneService<S> {
-    fn prune_to_height(
-        &self,
-        requested_height: u32,
-    ) -> core::result::Result<PruneResult, PruneServiceError> {
-        let policy = PrunePolicy {
-            target_size_mb: 0,
-            keep_below_tip: CORE_REORG_SAFETY_MARGIN,
-        };
-        let authority = self
-            .authority
-            .begin()
-            .map_err(|error| PruneServiceError::failed(error.to_string()))?;
-        let applied_tip_height = authority
-            .applied_tip_height()
-            .ok_or_else(|| PruneServiceError::failed("applied tip is unavailable"))?;
-        let mut pruneheight = self.pruneheight.lock();
-        let updated_pruneheight =
-            pruneheight.map_or(requested_height, |height| height.max(requested_height));
-        let safe_prune_height = applied_tip_height.saturating_sub(policy.retention_depth());
-        if updated_pruneheight > safe_prune_height {
-            return Err(PruneServiceError::failed(
-                "prune height is within reorg safety margin",
-            ));
-        }
-        let pruner_tip = updated_pruneheight
-            .checked_add(policy.retention_depth())
-            .ok_or_else(|| PruneServiceError::failed("prune height overflow"))?;
-
-        let prune_candidates: Vec<(u32, bitcoin_rs_primitives::BlockHash, usize)> = {
-            let blocks = self.blocks.read();
-            blocks
-                .iter()
-                .filter(|record| record.height < updated_pruneheight && record.tx_count > 0)
-                .map(|record| (record.height, record.hash, record.tx_count))
-                .collect()
-        };
-
-        let mut pruned_txids = Vec::new();
-        for (height, hash, tx_count) in prune_candidates {
-            if tx_count == 0 {
-                continue;
-            }
-            let bytes = self
-                .block_body_store
-                .load_block_body(height, hash.0)
-                .map_err(|error| PruneServiceError::failed(error.to_string()))?
-                .unwrap_or_default();
-            if bytes.is_empty() {
-                continue;
-            }
-            let block = deserialize::<Block>(&bytes).map_err(|error| {
-                PruneServiceError::failed(format!(
-                    "stored block body at height {height} failed decode: {error}"
-                ))
-            })?;
-            pruned_txids.extend(block.txs.iter().map(Tx::txid));
-        }
-        let mut batch = self.store.new_batch();
-        let (block_outcome, undo_outcome, prunable_files) = stage_block_and_undo_prune(
-            &*self.store,
-            &mut batch,
-            &self.block_files,
-            pruner_tip,
-            self.durable_tip_height.load(Ordering::Acquire),
-            policy,
-        )
-        .map_err(|err| PruneServiceError::failed(err.to_string()))?;
-        batch.put(
-            ColumnFamily::UtxoMeta,
-            PRUNEHEIGHT_METADATA_KEY,
-            &updated_pruneheight.to_be_bytes(),
-        );
-        self.store
-            .write(batch)
-            .map_err(|err| PruneServiceError::failed(err.to_string()))?;
-        reclaim_staged_flat_block_files(&*self.store, &self.block_files, &prunable_files)
-            .map_err(|err| PruneServiceError::failed(err.to_string()))?;
-
-        if !pruned_txids.is_empty() {
-            let mut transactions = self.transactions.write();
-            for txid in pruned_txids {
-                transactions.remove(&txid);
-            }
-        }
-
-        *pruneheight = Some(updated_pruneheight);
-
-        Ok(PruneResult {
-            requested_height,
-            pruneheight: updated_pruneheight,
-            block_rows_removed: block_outcome.blocks_removed,
-            undo_rows_removed: undo_outcome.blocks_removed,
-            bytes_freed: block_outcome
-                .bytes_freed
-                .saturating_add(undo_outcome.bytes_freed),
-        })
-    }
-
-    fn status(&self) -> PruneStatus {
-        PruneStatus {
-            pruned: true,
-            pruneheight: *self.pruneheight.lock(),
-        }
-    }
-}
-
-fn tx_index_capabilities(config: &NodeConfig) -> bitcoin_rs_index::IndexCapabilities {
-    bitcoin_rs_index::IndexCapabilities {
-        // Full ScriptIndex-backed Esplora responses need exact historical
-        // transactions to render prevouts and calculate fees. `utxo` owns
-        // only the compact live-output view and must not pay for TxLookup.
-        // `tx_index_query` still exposes TxLookup to Core RPCs only for an
-        // explicit --txindex configuration.
-        tx_lookup: config.indexes.txindex || config.indexes.script_index.keeps_history(),
-        script_history: config.indexes.script_index.keeps_history(),
-        script_live: config.indexes.script_index.is_enabled(),
-    }
-}
-
-fn build_tx_index_open_spec(
-    config: &NodeConfig,
-    txindex_cache_bytes: u64,
-    epoch: u64,
-) -> Result<Option<crate::txindex_worker::TxIndexOpenSpec>> {
-    let enabled = tx_index_capabilities(config);
-    if enabled.is_empty() {
-        return Ok(None);
-    }
-    if config.storage.prune_target_mb > 0 {
-        bail!("transaction and script indexing are not compatible with -prune");
-    }
-    let canonical_data_root = config
-        .data_dir
-        .canonicalize()
-        .unwrap_or_else(|_| config.data_dir.clone());
-    Ok(Some(crate::txindex_worker::TxIndexOpenSpec {
-        data_dir: config.data_dir.clone(),
-        namespace: "txindex",
-        storage_backend: config.storage.backend,
-        cache_bytes: txindex_cache_bytes,
-        epoch,
-        enabled,
-        rollback_rebuild_cutover: crate::txindex_worker::DEFAULT_ROLLBACK_REBUILD_CUTOVER,
-        canonical_data_root,
-        utxo: None,
-        chain_transition: None,
-    }))
-}
-
-struct TxIndexSpawn {
-    spec: crate::txindex_worker::TxIndexOpenSpec,
-    generation: crate::txindex_worker::Generation,
-    block_source: crate::txindex_worker::IndexBlockSource,
-    body_source: Arc<dyn BlockBodySource>,
-    wake_rx: Receiver<()>,
-    recovery_reporter: Arc<crate::recovery_evidence::RecoveryReporter>,
-}
 
 /// Aggregate handle to a running node.
 pub struct NodeState {
@@ -1195,480 +140,22 @@ pub struct NodeState {
     warning_store: Arc<crate::recovery_evidence::WarningStore>,
 }
 
-impl NodeState {
-    /// Opens (or creates) the node's data directory and configured storage
-    /// backend.
-    /// Derived-index workers are constructed dormant (`Opening`) and started
-    /// by [`Self::start_index_workers`] once crash recovery has made the
-    /// applied tip authoritative; `start_node` performs both steps.
-    #[allow(clippy::arc_with_non_send_sync)]
-    #[allow(clippy::too_many_lines)]
-    pub fn open(
-        config: NodeConfig,
-        mempool_observer: Option<&Arc<dyn bitcoin_rs_mempool::MempoolObserver>>,
-    ) -> Result<Self> {
-        config.validate()?;
-        std::fs::create_dir_all(&config.data_dir)
-            .with_context(|| format!("create data_dir {}", config.data_dir.display()))?;
-        let checkpoint_data_dir = crate::checkpoint_fs::open_data_dir(&config.data_dir)
-            .with_context(|| format!("open data_dir {}", config.data_dir.display()))?;
-        crate::checkpoint_fs::ensure_current_schema(&checkpoint_data_dir).with_context(|| {
-            format!(
-                "validate CURRENT_SCHEMA for datadir {}",
-                config.data_dir.display()
-            )
-        })?;
-        // Allocate the process epoch before anything else can consume one:
-        // durable, strictly greater than every earlier run of this data dir.
-        let epoch = allocate_process_epoch(&checkpoint_data_dir)?;
-        let checkpoint_config = crate::checkpoint::HeaderCheckpointConfig {
-            network: config.network,
-            genesis: config.network.genesis_block_hash(),
-        };
-        let checkpoint_load =
-            crate::checkpoint::load_checkpoint_from_dir(&checkpoint_data_dir, checkpoint_config)?;
-
-        // Divide the process cache budget across the persistent namespaces
-        // that exist in this deployment. A disabled txindex share redistributes
-        // to chainstate.
-        let cache_budget = bitcoin_rs_storage::clamp_dbcache_bytes(config.storage.dbcache_mb);
-        let cache_shares = bitcoin_rs_storage::split_cache_budget(
-            cache_budget,
-            !tx_index_capabilities(&config).is_empty(),
-        );
-        let chainstate_cache_bytes = cache_shares[0].bytes;
-        let txindex_cache_bytes = cache_shares[1].bytes;
-        let block_files =
-            Arc::new(FlatFileBlockStore::open(&config.data_dir).map_err(anyhow::Error::new)?);
-        let storage = NodeStorage::open(&config, chainstate_cache_bytes, Arc::clone(&block_files))?;
-        let undo_store = storage.undo_store();
-        // Before anything reads the chainstate, let alone serves or syncs it.
-        // A node that starts on a torn chainstate builds on it, and every block
-        // it adds makes the damage harder to find.
-        if let Some(marker) = undo_store
-            .load_disconnect_marker()
-            .map_err(anyhow::Error::new)?
-        {
-            let force_full_revalidation = requires_full_revalidation(&config.data_dir);
-            if marker.phase == crate::apply::DisconnectPhase::RolledBack && force_full_revalidation
-            {
-                undo_store.disarm_disconnect().map_err(anyhow::Error::new)?;
-                tracing::warn!(
-                    height = marker.height,
-                    hash = %marker.hash,
-                    "accepting completed deep reorg; full chain validation is required"
-                );
-            } else {
-                // Names directories rather than a `-reindex` option, because this
-                // node has no reindex. An instruction the operator cannot follow is
-                // worse than none.
-                //
-                // Remove the authoritative views. The marker covers a disconnect
-                // that did not reach a clean UTXO-and-tip checkpoint. TxIndex rows
-                // are derived state outside this marker, but a retained TxIndex
-                // watermark can stall rollback because wiping the chainstate
-                // removes the body positions the index refers to. Include the txindex
-                // path so the operator action is complete.
-                bail!(
-                    "refusing to start: a disconnect of block {hash} at height {height} did not \
-                     reach a clean checkpoint, so the UTXO set and chain tip cannot be trusted \
-                     together. The node cannot repair this in place. Remove or quarantine \
-                     {chainstate}, {checkpoints}, and {txindex}, then resync.",
-                    hash = marker.hash,
-                    height = marker.height,
-                    chainstate = config.data_dir.join("chainstate").display(),
-                    checkpoints = config.data_dir.join("chainstate-checkpoints").display(),
-                    txindex = config.data_dir.join("txindex").display(),
-                );
-            }
+impl Drop for NodeState {
+    fn drop(&mut self) {
+        let _admission = self.apply_handles.admission.close();
+        // Safety net: if `bounded_index_shutdown` was not called (e.g. in
+        // tests that drop `NodeState` directly), request shutdown and join
+        // any worker not already taken by `bounded_index_shutdown`.
+        if let Some(runtime) = &self.tx_index_runtime {
+            runtime.request_shutdown();
         }
-        let block_body_store = storage.block_body_store();
-
-        let zmq_endpoints = config.zmq_endpoints();
-        #[cfg(feature = "zmq")]
-        let zmq_publisher: Arc<dyn crate::ZmqPublisher> = if zmq_endpoints.is_empty() {
-            Arc::new(crate::NoOpZmqPublisher)
-        } else {
-            Arc::new(crate::SocketZmqPublisher::bind(zmq_endpoints)?)
-        };
-        #[cfg(not(feature = "zmq"))]
-        let zmq_publisher: Arc<dyn crate::ZmqPublisher> = {
-            let _ = &zmq_endpoints;
-            Arc::new(crate::NoOpZmqPublisher)
-        };
-        let InitialChainstate {
-            utxo: mut utxo_set,
-            coin_stats: initial_coin_stats,
-            tree: block_tree_value,
-            applied_tip: restored_applied_tip,
-            chain_tx_count: restored_chain_tx_count,
-            resume_source,
-            journal_bootstrap,
-        } = prepare_initial_chainstate(
-            checkpoint_load,
-            &checkpoint_data_dir,
-            checkpoint_config,
-            &config,
-        )?;
-        if resume_source == ResumeSource::Checkpoint {
-            tracing::info!(
-                height = restored_applied_tip.as_ref().map_or(0, |tip| tip.height),
-                hash = %restored_applied_tip
-                    .as_ref()
-                    .map_or_else(|| config.network.genesis_block_hash(), |tip| tip.hash),
-                "restored chainstate checkpoint"
-            );
+        if let Some(worker) = self.tx_index_worker.take() {
+            worker.join();
         }
-        // A2: Create the process-wide rollback-evidence warning store before
-        // any detection or worker spawn. One ArcSwap holds the complete
-        // immutable snapshot; getblockchaininfo loads one per request.
-        let warning_store = Arc::new(crate::recovery_evidence::WarningStore::new());
-        // A2: Read the durable applied-tip witness and detect checkpoint
-        // fallback. Emit a structured WARN, update the warning snapshot, and
-        // durably publish the event marker — only after all conditions hold:
-        // valid format/bounds, matching genesis, older writer epoch, and
-        // strictly greater witness height than the restored tip.
-        let genesis_hex = config.network.genesis_block_hash().to_string_be();
-        // One reporter routes every rollback fact of this process — the
-        // checkpoint fallback detected here and the index-ahead rewinds the
-        // txindex worker detects later — through the same warning store and
-        // event marker.
-        let recovery_reporter = Arc::new(crate::recovery_evidence::RecoveryReporter::new(
-            Arc::clone(&warning_store),
-            config.data_dir.clone(),
-            genesis_hex.clone(),
-            epoch,
-        ));
-        let restored_height = restored_applied_tip.as_ref().map_or(0, |tip| tip.height);
-        let restored_hash = restored_applied_tip
-            .as_ref()
-            .map_or_else(|| config.network.genesis_block_hash(), |tip| tip.hash)
-            .to_string_be();
-        if let Some(witness) =
-            crate::recovery_evidence::read_witness(&config.data_dir, &genesis_hex)
-        {
-            if let Some((witness_height, _)) = crate::recovery_evidence::detect_checkpoint_fallback(
-                &witness,
-                epoch,
-                &genesis_hex,
-                restored_height,
-            ) {
-                let source = match resume_source {
-                    ResumeSource::Cold => "cold",
-                    ResumeSource::Checkpoint => "checkpoint",
-                    ResumeSource::Journal => "journal",
-                };
-                recovery_reporter
-                    .report_checkpoint_fallback(
-                        witness_height,
-                        restored_height,
-                        &restored_hash,
-                        source,
-                        &witness.block_hash,
-                        std::time::SystemTime::now()
-                            .duration_since(std::time::UNIX_EPOCH)
-                            .map_or(0, |d| d.as_secs()),
-                    )
-                    .context("write checkpoint-fallback event marker")?;
-                let gap = witness_height.saturating_sub(restored_height);
-                if gap > STALE_RESTORE_ERROR_THRESHOLD {
-                    tracing::error!(
-                        witness_height,
-                        restored_height,
-                        gap,
-                        threshold = STALE_RESTORE_ERROR_THRESHOLD,
-                        "stale checkpoint restore: chainstate is {gap} blocks behind \
-                         the last durable applied-tip witness — no committed journal \
-                         suffix covers the gap, so the node is proceeding with a valid \
-                         but far-behind tip and the sync layer must re-fetch it"
-                    );
-                }
-            }
-        }
-        // Anchor the initial snapshot before `restored_applied_tip` is moved
-        // into the applied-tip slot: a restored node resumes at its restored
-        // tip, a fresh one at genesis, both with an untouched sequence.
-        let initial_snapshot = ChainSnapshot {
-            epoch,
-            sequence: 0,
-            tip_hash: restored_applied_tip
-                .as_ref()
-                .map_or_else(|| config.network.genesis_block_hash(), |tip| tip.hash),
-            tip_height: restored_applied_tip.as_ref().map_or(0, |tip| tip.height),
-        };
-        let coin_stats_listener =
-            bitcoin_rs_utxo::stats::CoinStatsListener::new(initial_coin_stats);
-        utxo_set.set_listener(Box::new(coin_stats_listener.clone()));
-        let journal = match journal_bootstrap {
-            Some(bootstrap) => {
-                Some(storage.journal_writer(open_journal_dir(&config.data_dir)?, bootstrap)?)
-            }
-            None => None,
-        };
-        let utxo = Arc::new(utxo_set);
-        let coin_stats = Arc::new(coin_stats_listener);
-        let mempool = Arc::new(RwLock::new(Mempool::new(MempoolLimits::default())));
-        let block_tree = Arc::new(RwLock::new(block_tree_value));
-        let chain_tip = block_tree.read().tip_handle();
-        let applied_tip: Arc<ArcSwapOption<TipSnapshot>> = Arc::new(ArcSwapOption::empty());
-        if let Some(restored_applied_tip) = restored_applied_tip {
-            applied_tip.store(Some(Arc::new(restored_applied_tip)));
-        }
-        let blocks = Arc::new(RwLock::new(BlockLog::new()));
-        let chain_tx_count = Arc::new(AtomicU64::new(restored_chain_tx_count));
-        let transactions = Arc::new(RwLock::new(HashMap::new()));
-        // Created before the txindex worker spawn: the worker mirrors this
-        // publisher's snapshot into its persisted consumer cursor.
-        let (chain_events_raw, chain_event_hints_rx_raw) =
-            ChainEventPublisher::new(epoch, initial_snapshot);
-        let shutdown = Arc::new(AtomicBool::new(false));
-        let chain_events = Arc::new(chain_events_raw);
-        let chain_transition = Arc::new(parking_lot::Mutex::new(()));
-        let tx_index_open_spec = build_tx_index_open_spec(&config, txindex_cache_bytes, epoch)?;
-        let (tx_index_runtime, tx_index_spawn, tx_index_lifecycle, tx_index_adapter) =
-            match tx_index_open_spec {
-                Some(mut spec) => {
-                    spec.utxo = Some(Arc::clone(&utxo));
-                    spec.chain_transition = Some(Arc::clone(&chain_transition));
-                    let (wake_tx, wake_rx) = crossbeam_channel::bounded(1);
-                    let runtime = Arc::new(crate::txindex_worker::TxIndexRuntime::new(wake_tx));
-                    let body_source: Arc<dyn BlockBodySource> =
-                        Arc::new(StoredBlockBodySource::new(Arc::clone(&block_body_store)));
-                    let block_source =
-                        crate::txindex_worker::IndexBlockSource::new(Arc::clone(&blocks))
-                            .with_block_body_source(Arc::clone(&body_source))
-                            .with_block_tree(Arc::clone(&block_tree));
-                    let lifecycle: Arc<arc_swap::ArcSwap<crate::txindex_worker::TxIndexLifecycle>> =
-                        Arc::new(arc_swap::ArcSwap::from_pointee(
-                            crate::txindex_worker::TxIndexLifecycle::Opening,
-                        ));
-                    let adapter = Arc::new(crate::txindex_worker::TxIndexQueryAdapter::new(
-                        Arc::clone(&lifecycle),
-                    ));
-                    let generation = crate::txindex_worker::Generation::new(spec.epoch);
-                    (
-                        Some(runtime),
-                        Some(TxIndexSpawn {
-                            spec,
-                            generation,
-                            block_source,
-                            body_source,
-                            wake_rx,
-                            recovery_reporter: Arc::clone(&recovery_reporter),
-                        }),
-                        Some(lifecycle),
-                        Some(adapter),
-                    )
-                }
-                None => (None, None, None, None),
-            };
-        let txindex_status = Arc::new(crate::txindex_worker::TxIndexCapability::new(
-            tx_index_lifecycle.clone(),
-            tx_index_runtime.clone(),
-            tx_index_capabilities(&config),
-        ));
-        let network = Arc::new(RwLock::new(NetworkState::default()));
-        let p2p = Arc::new(bitcoin_rs_p2p::P2pService::new(
-            bitcoin_rs_p2p::P2pServiceConfig {
-                listen_addrs: config.p2p.listen.clone(),
-                magic: bitcoin::p2p::Magic::from_bytes(config.p2p.magic),
-                dns_seeds_enabled: config.p2p.dns_seeds_enabled,
-                dns_seeds: config
-                    .network
-                    .dns_seeds()
-                    .iter()
-                    .map(|seed| (*seed).to_owned())
-                    .collect(),
-                dns_port: config.network.default_p2p_port(),
-                fixed_peers: config.p2p.connect.clone(),
-                outbound_active_limit: P2P_OUTBOUND_QUEUE_LIMIT,
-                outbound_peer_target: P2P_OUTBOUND_QUEUE_LIMIT,
-                outbound_queue_limit: P2P_OUTBOUND_QUEUE_LIMIT,
-                inbound_block_queue_limit: INBOUND_BLOCK_CHANNEL_LIMIT,
-                download_budget: bitcoin_rs_p2p::default_sync_budget(),
-            },
-            Arc::clone(&shutdown),
-        ));
-        let network_active = p2p.network_active_handle();
-        let banned = p2p.banned_handle();
-        let peer_table = p2p.table();
-        let p2p_outbound_tx = p2p.outbound_sender();
-        let p2p_outbound_rx = p2p.outbound_receiver();
-        let inbound_headers_tx = p2p.inbound_headers_sender();
-        let inbound_headers_rx = p2p.inbound_headers_receiver();
-        let inbound_blocks_tx = p2p.inbound_blocks_sender();
-        let inbound_blocks_rx = p2p.inbound_blocks_receiver();
-        let (inbound_tx_tx, inbound_tx_rx_raw) =
-            crossbeam_channel::bounded::<bitcoin_rs_p2p::InboundTx>(INBOUND_TX_CHANNEL_LIMIT);
-        let inbound_tx_rx = Arc::new(Mutex::new(inbound_tx_rx_raw));
-        let chain_event_hints_rx = Arc::new(Mutex::new(chain_event_hints_rx_raw));
-        // The template-coordinator wake exists from node birth so the apply
-        // path and the gateway can fire it before `run` builds the
-        // coordinator; the coordinator attaches itself once constructed.
-        let mining_generation = Arc::new(crate::mining::MiningGenerationSignal::new());
-        // One gateway per pool. Mempool owns fan-out: mining occupies the
-        // observer slot, and ZMQ sequence (or a test observer) attaches as an
-        // extra named leg. Admission/relay legs attach after construction.
-        let mempool_gateway = {
-            let publisher = Arc::clone(&zmq_publisher);
-            let cloned_mining = Arc::clone(&mining_generation);
-            let mining_leg: Arc<dyn bitcoin_rs_mempool::MempoolObserver> = cloned_mining;
-            let gateway =
-                bitcoin_rs_mempool::MempoolGateway::shared_with(Arc::clone(&mempool), mining_leg);
-            if publisher.wants_notifications() {
-                gateway
-                    .attach_observer_leg(
-                        "sequence",
-                        Arc::new(bitcoin_rs_rpc::zmq::MempoolSequenceObserver::new(publisher)),
-                    )
-                    .map_err(anyhow::Error::msg)?;
-            } else if let Some(observer) = mempool_observer.cloned() {
-                gateway
-                    .attach_observer_leg("test", observer)
-                    .map_err(anyhow::Error::msg)?;
-            }
-            gateway
-        };
-        // Construct followers before Chainstate so capture policy has one owner.
-        let followers = crate::chain_effects::ChainFollowers::new(
-            crate::chain_effects::ChainEffects::new(
-                Arc::clone(&blocks),
-                Arc::clone(&zmq_publisher),
-                tx_index_runtime.clone(),
-            ),
-            Arc::clone(&mining_generation),
-            Some(Arc::clone(&mempool_gateway)),
-        );
-        let (capture_rawtx, capture_block_bytes) = followers.capture_flags();
-        let mut apply_handles = crate::apply::Chainstate {
-            network: config.network,
-            chain_tip: Arc::clone(&chain_tip),
-            applied_tip: Arc::clone(&applied_tip),
-            chain_tx_count: Arc::clone(&chain_tx_count),
-            applied_seq: Arc::new(AtomicU64::new(0)),
-            block_tree: Arc::clone(&block_tree),
-            utxo: Arc::clone(&utxo),
-            coin_stats: Arc::clone(&coin_stats),
-            mempool: Arc::clone(&mempool),
-            mempool_gateway: Arc::clone(&mempool_gateway),
-            chain_events: Arc::clone(&chain_events),
-            block_body_store: Some(Arc::clone(&block_body_store)),
-            undo_store,
-            admission: Arc::new(crate::apply::ApplyAdmission::new()),
-            shutdown: Arc::clone(&shutdown),
-            chain_transition,
-            assume_valid_height: config.validation.assume_valid_height,
-            assume_valid_gate: Arc::new(crate::apply::AssumeValidGate::new(
-                config.network,
-                config.validation.assume_valid_height,
-            )),
-            journal,
-            checkpoint_publisher: None,
-            capture_rawtx,
-            capture_block_bytes,
-        };
-        apply_handles.assume_valid_gate.evaluate(&block_tree.read());
-        // A restored checkpoint is durable at its own height by definition, so
-        // start there rather than at zero, which would refuse all undo pruning.
-        let durable_tip_height = Arc::new(AtomicU32::new(
-            applied_tip.load().as_ref().map_or(0, |tip| tip.height),
-        ));
-        apply_handles.checkpoint_publisher =
-            Some(Arc::new(crate::checkpoint_worker::CheckpointPublisher {
-                admission: Arc::clone(&apply_handles.admission),
-                undo_store: Arc::clone(&apply_handles.undo_store),
-                block_body_store: Arc::clone(&block_body_store),
-                applied_tip: Arc::clone(&applied_tip),
-                checkpoint_data_dir: crate::checkpoint_fs::open_data_dir(&config.data_dir)
-                    .with_context(|| format!("open data_dir {}", config.data_dir.display()))?,
-                network: config.network,
-                genesis_hash: config.network.genesis_block_hash(),
-                block_tree: Arc::clone(&block_tree),
-                utxo: Arc::clone(&utxo),
-                coin_stats: Arc::clone(&coin_stats),
-                chain_tx_count: Arc::clone(&chain_tx_count),
-                journal: apply_handles.journal.clone(),
-                data_dir: config.data_dir.clone(),
-                chain_events: Arc::clone(&chain_events),
-                durable_tip_height: Arc::clone(&durable_tip_height),
-            }));
-        let sync = Arc::new(crate::BlockSync::new(
-            apply_handles.clone(),
-            followers.clone(),
-            Arc::clone(&peer_table),
-            Arc::clone(&inbound_headers_rx),
-            Arc::clone(&inbound_blocks_rx),
-        ));
-        let prune_service = if config.storage.prune_target_mb > 0 {
-            Some(storage.prune_service(
-                &block_files,
-                &block_body_store,
-                Arc::clone(&blocks),
-                Arc::clone(&transactions),
-                apply_handles.prune_authority(),
-                &durable_tip_height,
-            )?)
-        } else {
-            None
-        };
-        tracing::info!(
-            backend = storage.kind(),
-            chainstate_dir = %config.data_dir.join("chainstate").display(),
-            chainstate_cache_bytes,
-            txindex_cache_bytes,
-            total_cache_bytes = cache_budget,
-            "opened storage backend with effective cache capacities"
-        );
-        let data_dir = config.data_dir.clone();
-        Ok(Self {
-            durable_tip_height,
-            config,
-            data_dir,
-            #[cfg(test)]
-            resume_source,
-            storage,
-            block_body_store,
-            utxo,
-            coin_stats,
-            tx_index_runtime,
-            tx_index_spawn,
-            tx_index_worker: None,
-            tx_index_lifecycle,
-            tx_index_adapter,
-            txindex_status,
-            prune_service,
-            zmq_publisher,
-            mempool,
-            mempool_gateway,
-            mining_generation,
-            chain_tip,
-            applied_tip,
-            chain_tx_count: Arc::clone(&chain_tx_count),
-            block_tree,
-            blocks,
-            transactions,
-            network,
-            network_active,
-            p2p,
-            peer_table,
-            banned,
-            p2p_outbound_tx,
-            p2p_outbound_rx,
-            inbound_headers_tx,
-            inbound_headers_rx,
-            inbound_blocks_tx,
-            inbound_blocks_rx,
-            inbound_tx_tx,
-            inbound_tx_rx,
-            chain_events: Arc::clone(&chain_events),
-            chain_event_hints_rx,
-            apply_handles,
-            followers,
-            sync,
-            warning_store,
-        })
     }
+}
 
+impl NodeState {
     /// Returns a borrow of the resolved configuration.
     #[must_use]
     pub const fn config(&self) -> &NodeConfig {
@@ -1684,83 +171,6 @@ impl NodeState {
     #[cfg(test)]
     pub(crate) const fn resume_source(&self) -> ResumeSource {
         self.resume_source
-    }
-
-    /// Publishes a durable clean checkpoint and returns the published
-    /// generation, or an error if there is no applied tip.
-    ///
-    /// This is the public boundary for the private checkpoint machinery; it
-    /// keeps `CheckpointWrite`, `CheckpointError`, and the checkpoint module
-    /// internal to the crate.
-    pub fn publish_checkpoint(&self) -> Result<u64> {
-        match self.write_clean_checkpoint()? {
-            crate::checkpoint::CheckpointWrite::SkippedNoAppliedTip => {
-                bail!("checkpoint refused: no applied tip to publish")
-            }
-            crate::checkpoint::CheckpointWrite::Published { generation } => Ok(generation),
-        }
-    }
-
-    /// Creates a [`crate::checkpoint_worker::CheckpointPublisher`] from this
-    /// state's shared handles, for use by the periodic checkpoint worker.
-    ///
-    /// The publisher owns its own `Dir` handle (reopened from the data-dir
-    /// path) and cloned `Arc`s, so it can be moved into a background thread
-    /// without borrowing from `self`.
-    pub(crate) fn checkpoint_publisher(
-        &self,
-    ) -> core::result::Result<
-        crate::checkpoint_worker::CheckpointPublisher,
-        crate::checkpoint::CheckpointError,
-    > {
-        Ok(crate::checkpoint_worker::CheckpointPublisher {
-            admission: Arc::clone(&self.apply_handles.admission),
-            undo_store: Arc::clone(&self.apply_handles.undo_store),
-            block_body_store: Arc::clone(&self.block_body_store),
-            applied_tip: Arc::clone(&self.applied_tip),
-            checkpoint_data_dir: crate::checkpoint_fs::open_data_dir(&self.data_dir)
-                .map_err(crate::checkpoint::CheckpointError::Io)?,
-            network: self.config.network,
-            genesis_hash: self.config.network.genesis_block_hash(),
-            block_tree: Arc::clone(&self.block_tree),
-            utxo: Arc::clone(&self.utxo),
-            coin_stats: Arc::clone(&self.coin_stats),
-            chain_tx_count: Arc::clone(&self.chain_tx_count),
-            journal: self.apply_handles.journal.clone(),
-            data_dir: self.data_dir.clone(),
-            chain_events: Arc::clone(&self.chain_events),
-            durable_tip_height: Arc::clone(&self.durable_tip_height),
-        })
-    }
-
-    /// Spawns the periodic checkpoint worker with a custom cadence.
-    ///
-    /// Production wiring goes through `start_node`, which uses
-    /// [`crate::checkpoint_worker::CHECKPOINT_INTERVAL_BLOCKS`] and
-    /// [`crate::checkpoint_worker::CHECKPOINT_INTERVAL_SECS`]. This method
-    /// is `pub` so integration tests can use a small cadence.
-    ///
-    /// Returns the worker's join handle. The worker exits when the node's
-    /// shutdown flag is set.
-    pub fn start_periodic_checkpoint(
-        &self,
-        interval_blocks: u32,
-        interval_secs: Duration,
-    ) -> Result<std::thread::JoinHandle<()>> {
-        let publisher = self.checkpoint_publisher().map_err(anyhow::Error::new)?;
-        Ok(crate::checkpoint_worker::spawn_periodic_checkpoint_worker(
-            publisher,
-            Arc::clone(&self.shutdown()),
-            interval_blocks,
-            interval_secs,
-        )?)
-    }
-
-    pub(crate) fn write_clean_checkpoint(
-        &self,
-    ) -> core::result::Result<crate::checkpoint::CheckpointWrite, crate::checkpoint::CheckpointError>
-    {
-        self.checkpoint_publisher()?.publish()
     }
 
     /// Returns the configured storage backend that was opened.
@@ -1779,83 +189,6 @@ impl NodeState {
     #[must_use]
     pub fn coin_stats(&self) -> Arc<bitcoin_rs_utxo::stats::CoinStatsListener> {
         Arc::clone(&self.coin_stats)
-    }
-
-    /// Returns the node-owned complete transaction-index query adapter.
-    #[must_use]
-    pub fn tx_index_query(&self) -> Option<Arc<dyn bitcoin_rs_rpc::context::TxIndexQuery>> {
-        if !self.config.indexes.txindex {
-            return None;
-        }
-        self.tx_index_adapter.as_ref().map(|adapter| {
-            let q: Arc<dyn bitcoin_rs_rpc::context::TxIndexQuery> = adapter.clone();
-            q
-        })
-    }
-
-    /// Returns transaction lookup for internal Esplora projections.
-    ///
-    /// `--scriptindex` builds this dependency as well, but that does not
-    /// enable or advertise the Core `--txindex` contract.
-    #[must_use]
-    pub fn esplora_tx_index_query(&self) -> Option<Arc<dyn bitcoin_rs_rpc::context::TxIndexQuery>> {
-        self.tx_index_adapter.as_ref().map(|adapter| {
-            let q: Arc<dyn bitcoin_rs_rpc::context::TxIndexQuery> = adapter.clone();
-            q
-        })
-    }
-
-    /// Returns the node-owned complete generic script-index query adapter.
-    #[must_use]
-    pub fn script_index_query(&self) -> Option<Arc<dyn bitcoin_rs_rpc::context::ScriptIndexQuery>> {
-        if !self.config.indexes.script_index.is_enabled() {
-            return None;
-        }
-        self.tx_index_adapter.as_ref().map(|adapter| {
-            let q: Arc<dyn bitcoin_rs_rpc::context::ScriptIndexQuery> = adapter.clone();
-            q
-        })
-    }
-
-    /// Starts the derived-index workers. Call only once the applied tip is
-    /// authoritative — after crash recovery — so the index reconciles against
-    /// the real chainstate and never mistakes a recovered gap for a stale branch.
-    pub fn start_index_workers(&mut self) -> anyhow::Result<()> {
-        let Some(spawn) = self.tx_index_spawn.take() else {
-            return Ok(());
-        };
-        let runtime = self
-            .tx_index_runtime
-            .as_ref()
-            .context("txindex runtime missing for a pending worker spawn")?;
-        let lifecycle = self
-            .tx_index_lifecycle
-            .as_ref()
-            .context("txindex lifecycle missing for a pending worker spawn")?;
-        let worker = crate::txindex_worker::TxIndexWorker::spawn_with_open(
-            Arc::clone(runtime),
-            spawn.spec,
-            Arc::clone(lifecycle),
-            spawn.generation,
-            Arc::clone(&self.applied_tip),
-            Arc::clone(&self.block_tree),
-            Some(Arc::clone(&self.block_body_store)),
-            spawn.block_source,
-            Some(spawn.body_source),
-            Arc::clone(&self.chain_events),
-            spawn.recovery_reporter,
-            Arc::clone(&self.apply_handles.shutdown),
-            spawn.wake_rx,
-        )
-        .context("spawn txindex worker")?;
-        self.tx_index_worker = Some(worker);
-        Ok(())
-    }
-
-    /// Returns the live txindex status source for `getcapabilities`.
-    #[must_use]
-    pub fn txindex_status(&self) -> Arc<dyn bitcoin_rs_rpc::capabilities::TxIndexCapabilitySource> {
-        self.txindex_status.clone()
     }
 
     /// Returns the manual pruning service when pruning is enabled.
@@ -2017,9 +350,6 @@ impl NodeState {
     /// Returns a cloned `Sender` that the P2P listener pushes inbound
     /// block headers into. The matching `Receiver` is polled by
     /// `BlockSync::tick` to extend the `BlockTree`.
-    /// Returns a cloned `Sender` that the P2P listener pushes inbound
-    /// block headers into. The matching `Receiver` is polled by
-    /// `BlockSync::tick` to extend the `BlockTree`.
     pub fn inbound_headers_sender(&self) -> Sender<bitcoin_rs_p2p::InboundHeaders> {
         self.inbound_headers_tx.clone()
     }
@@ -2094,56 +424,6 @@ impl NodeState {
         Arc::clone(&self.apply_handles.shutdown)
     }
 
-    /// Bounded txindex-worker shutdown: requests the worker shutdown, waits up
-    /// to `deadline` for a clean join, and detaches on
-    /// expiry. On detach, revokes the generation token and publishes
-    /// `ShutdownAbandoned` so queries return typed `Unavailable` instead of
-    /// hitting a torn reader.
-    pub(crate) fn bounded_index_shutdown(&mut self, deadline: Duration) {
-        let start = std::time::Instant::now();
-        if let Some(runtime) = &self.tx_index_runtime {
-            runtime.request_shutdown();
-        }
-        // Take the worker out of self so we can join it without holding self
-        // mutably across the wait.
-        let tx_index_worker = self.tx_index_worker.take();
-        let tx_deadline = start + deadline;
-        let tx_joined = if let Some(mut worker) = tx_index_worker {
-            let mut joined = false;
-            while std::time::Instant::now() < tx_deadline {
-                if worker.is_finished() {
-                    break;
-                }
-                std::thread::sleep(Duration::from_millis(10));
-            }
-            if worker.is_finished() {
-                worker.join();
-                joined = true;
-            } else {
-                tracing::warn!("txindex worker still blocked; abandoning join");
-                // Revoke the generation token so late publication is a no-op.
-                if let Some(generation_token) = &worker.generation {
-                    generation_token.revoke();
-                }
-                if let Some(lifecycle) = &self.tx_index_lifecycle {
-                    lifecycle.store(Arc::new(
-                        crate::txindex_worker::TxIndexLifecycle::ShutdownAbandoned,
-                    ));
-                }
-                // Poison the namespace so it cannot be reclaimed in this process.
-                worker.poison_namespace();
-                // Detach the join handle so Drop does not block on join.
-                // The worker thread continues running but will exit after
-                // shutdown is observed; Drop is a no-op for the handle.
-                worker.detach();
-            }
-            joined
-        } else {
-            true
-        };
-        let _ = tx_joined;
-    }
-
     /// Clone of the chainstate facade used by apply, reorg, and sync.
     #[must_use]
     pub fn chainstate(&self) -> crate::apply::Chainstate {
@@ -2174,2599 +454,5 @@ impl NodeState {
     }
 }
 
-impl Drop for NodeState {
-    fn drop(&mut self) {
-        let _admission = self.apply_handles.admission.close();
-        // Safety net: if `bounded_index_shutdown` was not called (e.g. in
-        // tests that drop `NodeState` directly), still request shutdown and
-        // `bounded_index_shutdown` took them.
-        if let Some(runtime) = &self.tx_index_runtime {
-            runtime.request_shutdown();
-        }
-        if let Some(worker) = self.tx_index_worker.take() {
-            worker.join();
-        }
-    }
-}
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use bitcoin_rs_index::IndexCapabilities;
-    use bitcoin_rs_primitives::Block;
-    use bitcoin_rs_primitives::BlockHash;
-    use bitcoin_rs_primitives::Hash256;
-    use bitcoin_rs_primitives::Header;
-    use bitcoin_rs_primitives::OutPoint;
-    use bitcoin_rs_primitives::Tx;
-    use bitcoin_rs_primitives::TxIn;
-    use bitcoin_rs_primitives::TxOut;
-    use bitcoin_rs_primitives::consensus_bytes;
-    use bitcoin_rs_primitives::encode::double_sha256;
-    use bitcoin_rs_rpc::context::BlockRecord;
-
-    /// IDX-01: scriptindex mode selects `ScriptLive` and/or `ScriptHistory`.
-    #[test]
-    fn script_index_capabilities_match_the_storage_contract() {
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.indexes.txindex = false;
-
-        config.indexes.script_index = crate::config::ScriptIndexMode::Disabled;
-        assert_eq!(tx_index_capabilities(&config), IndexCapabilities::NONE);
-
-        config.indexes.script_index = crate::config::ScriptIndexMode::Utxo;
-        assert_eq!(
-            tx_index_capabilities(&config),
-            IndexCapabilities::SCRIPT_LIVE,
-            "utxo mode owns only the compact live-output view"
-        );
-
-        config.indexes.script_index = crate::config::ScriptIndexMode::Full;
-        assert_eq!(tx_index_capabilities(&config), IndexCapabilities::ALL);
-
-        config.indexes.txindex = true;
-        config.indexes.script_index = crate::config::ScriptIndexMode::Disabled;
-        assert_eq!(tx_index_capabilities(&config), IndexCapabilities::TX_LOOKUP);
-
-        config.indexes.script_index = crate::config::ScriptIndexMode::Utxo;
-        assert_eq!(
-            tx_index_capabilities(&config),
-            IndexCapabilities {
-                tx_lookup: true,
-                script_history: false,
-                script_live: true,
-            }
-        );
-    }
-
-    fn publish_applied_tip_height(state: &NodeState, height: u32) {
-        let mut hash = [0_u8; 32];
-        hash[..size_of::<u32>()].copy_from_slice(&height.to_le_bytes());
-        state.applied_tip.store(Some(Arc::new(TipSnapshot {
-            tip_id: bitcoin_rs_chain::node::NodeId::new(height),
-            height,
-            chainwork: bitcoin_rs_chain::node::ChainWork::ZERO,
-            hash: bitcoin_rs_primitives::Hash256::from_le_bytes(&hash),
-        })));
-    }
-
-    #[test]
-    fn full_revalidation_marker_is_sticky_when_journal_is_disabled() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.chainstate_journal.enabled = false;
-        let journal_dir = config.data_dir.join(CHAINSTATE_JOURNAL_DIR);
-        let marker = journal_dir.join(crate::chainstate_journal::FULL_REVALIDATION_MARKER);
-        std::fs::create_dir_all(&journal_dir)?;
-        std::fs::write(&marker, b"force full validation\n")?;
-
-        assert!(requires_full_revalidation(&config.data_dir));
-        Ok(())
-    }
-
-    #[test]
-    fn open_constructs_empty_handles() -> anyhow::Result<()> {
-        use tempfile::tempdir;
-
-        let dir = tempdir()?;
-        let config = crate::NodeConfig {
-            data_dir: dir.path().join("node"),
-            ..crate::NodeConfig::default_for_network(crate::Network::Regtest)
-        };
-
-        let state = NodeState::open(config, None)?;
-        let utxo = state.utxo();
-        let mempool = state.mempool();
-
-        assert!(
-            Arc::strong_count(&utxo) >= 2,
-            "caller and NodeState should both hold a strong ref"
-        );
-        assert!(Arc::strong_count(&mempool) >= 2);
-        assert_eq!(mempool.read().len(), 0, "fresh mempool must be empty");
-
-        Ok(())
-    }
-
-    #[test]
-    fn open_constructs_empty_block_tree() -> anyhow::Result<()> {
-        use tempfile::tempdir;
-
-        let dir = tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-        let tree = state.block_tree();
-
-        assert!(
-            tree.read().is_empty(),
-            "freshly opened tree has zero headers"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn open_constructs_coin_stats_listener() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-        let snapshot = state.coin_stats().snapshot();
-        assert_eq!(
-            snapshot.tx_count, 0,
-            "freshly opened coin_stats has zero txs"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn open_skips_tx_index_when_disabled() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-
-        assert!(
-            state.tx_index_query().is_none(),
-            "txindex disabled by default"
-        );
-        assert!(
-            !state.data_dir().join("txindex").exists(),
-            "disabled txindex must not create storage"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn open_constructs_tx_index_when_enabled() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.indexes.txindex = true;
-        let mut state = NodeState::open(config, None)?;
-        state.start_index_workers()?;
-        let (Some(a), Some(b)) = (state.tx_index_query(), state.tx_index_query()) else {
-            panic!("txindex query engine missing when enabled");
-        };
-        assert!(Arc::ptr_eq(&a, &b), "txindex query handle must be stable");
-        // The worker opens the store asynchronously; the directory appears
-        // once its open completes, not during NodeState::open.
-        let deadline = std::time::Instant::now() + Duration::from_secs(30);
-        while !state.data_dir().join("txindex").exists() {
-            assert!(
-                std::time::Instant::now() < deadline,
-                "enabled txindex worker did not create storage within 30s"
-            );
-            std::thread::sleep(Duration::from_millis(20));
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn index_workers_start_only_when_asked() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.indexes.txindex = true;
-        let mut state = NodeState::open(config, None)?;
-
-        assert!(state.tx_index_lifecycle.as_ref().is_some_and(|lifecycle| {
-            matches!(
-                lifecycle.load().as_ref(),
-                crate::txindex_worker::TxIndexLifecycle::Opening
-            )
-        }));
-        assert!(state.tx_index_worker.is_none());
-
-        state.start_index_workers()?;
-        assert!(state.tx_index_worker.is_some());
-        let deadline = std::time::Instant::now() + Duration::from_secs(10);
-        while state.tx_index_lifecycle.as_ref().is_some_and(|lifecycle| {
-            matches!(
-                lifecycle.load().as_ref(),
-                crate::txindex_worker::TxIndexLifecycle::Opening
-            )
-        }) {
-            assert!(
-                std::time::Instant::now() < deadline,
-                "txindex lifecycle remained Opening"
-            );
-            std::thread::sleep(Duration::from_millis(20));
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn script_index_builds_without_advertising_core_txindex() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.indexes.txindex = false;
-        config.indexes.script_index = crate::config::ScriptIndexMode::Full;
-
-        let mut state = NodeState::open(config, None)?;
-        state.start_index_workers()?;
-
-        assert!(state.chain_followers().effects().tx_index().is_some());
-        assert!(state.tx_index_query().is_none());
-        assert!(state.esplora_tx_index_query().is_some());
-        assert!(state.script_index_query().is_some());
-        // The script-index worker shares the txindex storage; it is created
-        // asynchronously once the worker's open completes.
-        let deadline = std::time::Instant::now() + Duration::from_secs(30);
-        while !state.data_dir().join("txindex").exists() {
-            assert!(
-                std::time::Instant::now() < deadline,
-                "script-index worker did not create storage within 30s"
-            );
-            std::thread::sleep(Duration::from_millis(20));
-        }
-        Ok(())
-    }
-
-    /// Opens a node in each accepted `scriptindex` mode and asserts the
-    /// concrete answer for `unspent_outputs`, rather than only the `full` path.
-    ///
-    /// `utxo` must be independently usable while historical rows are absent;
-    /// this test pins that both enabled modes converge to a concrete answer.
-    #[test]
-    fn script_index_modes_give_concrete_unspent_outputs_answers() -> anyhow::Result<()> {
-        use bitcoin_rs_index::ScriptHash;
-        use bitcoin_rs_rpc::context::TxQueryError;
-
-        // Genesis is applied in each case so the index worker has a published
-        // chain tip to anchor. The disabled case has no adapter; the enabled
-        // modes must publish their own capability watermarks independently.
-        let scripthash = ScriptHash::from_script_bytes(&[0x51, 0x01]);
-
-        // `disabled`: no script index at all, so no query adapter is handed
-        // out. The answer is a definite "not available", not a retry.
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.indexes.txindex = false;
-        config.indexes.script_index = crate::config::ScriptIndexMode::Disabled;
-        let state = NodeState::open(config, None)?;
-        let _ = state.apply_block(&crate::Network::Regtest.genesis_block())?;
-        assert!(
-            state.script_index_query().is_none(),
-            "disabled must not hand out a script-index query adapter"
-        );
-        drop(state);
-
-        // `utxo`: the compact live-output view is independently usable while
-        // historical script rows are not maintained.
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.indexes.txindex = false;
-        config.indexes.script_index = crate::config::ScriptIndexMode::Utxo;
-        let mut state = NodeState::open(config, None)?;
-        state.start_index_workers()?;
-        let _ = state.apply_block(&crate::Network::Regtest.genesis_block())?;
-        let Some(query) = state.script_index_query() else {
-            panic!("utxo mode must hand out a script-index query adapter")
-        };
-        let deadline = std::time::Instant::now() + Duration::from_secs(30);
-        loop {
-            match query.unspent_outputs(scripthash) {
-                Ok(records) => {
-                    assert!(records.is_empty(), "an unfunded script has no outputs");
-                    match query.history_snapshot(scripthash) {
-                        Err(TxQueryError::Unavailable(reason)) => {
-                            assert!(
-                                reason.contains("script history is disabled"),
-                                "utxo history must be disabled, not lagging: {reason}"
-                            );
-                        }
-                        other => panic!("utxo mode must fail history as disabled, got {other:?}"),
-                    }
-                    break;
-                }
-                Err(TxQueryError::Retry | TxQueryError::Unavailable(_)) => {
-                    assert!(
-                        std::time::Instant::now() < deadline,
-                        "utxo mode must converge on the live view"
-                    );
-                    std::thread::sleep(Duration::from_millis(20));
-                }
-                Err(error) => panic!("unexpected script-index error: {error}"),
-            }
-        }
-
-        // `full`: the accepted mode. It converges on a concrete answer — an
-        // empty set for an unfunded script — rather than retrying forever.
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.indexes.txindex = false;
-        config.indexes.script_index = crate::config::ScriptIndexMode::Full;
-        let mut state = NodeState::open(config, None)?;
-        state.start_index_workers()?;
-        let _ = state.apply_block(&crate::Network::Regtest.genesis_block())?;
-        let Some(query) = state.script_index_query() else {
-            panic!("full mode must hand out a script-index query adapter")
-        };
-
-        let deadline = std::time::Instant::now() + Duration::from_secs(30);
-        loop {
-            match query.unspent_outputs(scripthash) {
-                Ok(records) => {
-                    assert!(
-                        records.is_empty(),
-                        "an unfunded script has no unspent outputs"
-                    );
-                    break;
-                }
-                // The worker indexes asynchronously, so it is legitimately
-                // still opening for a bounded interval.
-                Err(TxQueryError::Retry | TxQueryError::Unavailable(_)) => {
-                    assert!(
-                        std::time::Instant::now() < deadline,
-                        "full mode must converge on a concrete answer, not retry forever"
-                    );
-                    std::thread::sleep(Duration::from_millis(20));
-                }
-                Err(error) => panic!("unexpected script-index error: {error}"),
-            }
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn drop_joins_txindex_worker_before_reopen() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.indexes.txindex = true;
-
-        {
-            let mut state = NodeState::open(config.clone(), None)?;
-            state.start_index_workers()?;
-            assert!(state.tx_index_query().is_some());
-        }
-
-        let reopened = NodeState::open(config, None)?;
-        assert!(reopened.tx_index_query().is_some());
-        Ok(())
-    }
-
-    #[test]
-    fn open_rejects_txindex_with_pruning() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.indexes.txindex = true;
-        config.storage.prune_target_mb = 1;
-
-        let error = match NodeState::open(config, None) {
-            Ok(_) => anyhow::bail!("txindex with pruning must be rejected"),
-            Err(error) => error,
-        };
-        assert!(
-            error
-                .to_string()
-                .contains("transaction and script indexing are not compatible with -prune"),
-            "unexpected error: {error:#}"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn open_constructs_block_sync_orchestrator() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-        let sync_a = state.sync();
-        let sync_b = state.sync();
-        assert!(
-            Arc::ptr_eq(&sync_a, &sync_b),
-            "sync handle is stable across calls"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn open_constructs_empty_applied_tip() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-
-        assert!(
-            state.applied_tip().load_full().is_none(),
-            "freshly opened applied_tip is empty"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn open_constructs_empty_peer_table() -> anyhow::Result<()> {
-        use tempfile::tempdir;
-
-        let dir = tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-
-        assert!(
-            state.peer_table().is_empty(),
-            "freshly opened table is empty"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn open_constructs_empty_peer_table_again() -> anyhow::Result<()> {
-        use tempfile::tempdir;
-
-        let dir = tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-
-        assert!(state.peer_table().is_empty());
-        Ok(())
-    }
-
-    #[test]
-    fn zmq_publisher_handle_defaults_to_noop() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-        let publisher = state.zmq_publisher();
-        // No-op publisher accepts publish calls silently.
-        publisher.publish_hashblock(bitcoin_rs_primitives::Hash256::default());
-        Ok(())
-    }
-
-    #[cfg(feature = "zmq")]
-    #[test]
-    fn zmq_publisher_handle_reports_active_metadata() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.notifications.zmq = vec![
-            bitcoin_rs_rpc::zmq::ZmqEndpointConfig {
-                endpoint: "inproc://state-zmq-block".to_owned(),
-                topics: vec![
-                    bitcoin_rs_rpc::zmq::ZmqTopic::HashBlock,
-                    bitcoin_rs_rpc::zmq::ZmqTopic::RawBlock,
-                ],
-                hwm: Some(17),
-            },
-            bitcoin_rs_rpc::zmq::ZmqEndpointConfig {
-                endpoint: "inproc://state-zmq-tx".to_owned(),
-                topics: vec![
-                    bitcoin_rs_rpc::zmq::ZmqTopic::HashTx,
-                    bitcoin_rs_rpc::zmq::ZmqTopic::RawTx,
-                ],
-                hwm: Some(20),
-            },
-        ];
-        let state = NodeState::open(config, None)?;
-
-        let notifications = state.zmq_publisher().active_notifiers();
-        let notification_types: Vec<_> = notifications
-            .iter()
-            .map(|notification| notification.topic.notifier_type())
-            .collect();
-        let hwms: Vec<_> = notifications
-            .iter()
-            .map(|notification| notification.hwm)
-            .collect();
-        assert_eq!(
-            notification_types,
-            ["pubhashblock", "pubrawblock", "pubhashtx", "pubrawtx"]
-        );
-        assert_eq!(hwms, [17, 17, 20, 20]);
-        Ok(())
-    }
-
-    #[test]
-    fn inbound_headers_sender_is_unbounded_clone_target() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-        let tx1 = state.inbound_headers_sender();
-        let tx2 = state.inbound_headers_sender();
-        tx1.send(bitcoin_rs_p2p::InboundHeaders {
-            headers: Vec::new(),
-            source: None,
-        })
-        .map_err(|err| anyhow::anyhow!("send via tx1 failed: {err}"))?;
-        tx2.send(bitcoin_rs_p2p::InboundHeaders {
-            headers: Vec::new(),
-            source: None,
-        })
-        .map_err(|err| anyhow::anyhow!("send via tx2 failed: {err}"))?;
-        Ok(())
-    }
-
-    #[test]
-    fn inbound_blocks_sender_is_clonable_into_listener_threads() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-        let _tx1 = state.inbound_blocks_sender();
-        let _tx2 = state.inbound_blocks_sender();
-        Ok(())
-    }
-
-    #[test]
-    fn inbound_blocks_channel_is_bounded_against_flood() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-        let tx = state.inbound_blocks_sender();
-        let block = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        // No `tick` drains the channel in this unit test, so it fills to the
-        // bound; the block past the limit must be rejected rather than queued
-        // (the OOM-flood guard), proving backpressure engages at the producer.
-        for _ in 0..super::INBOUND_BLOCK_CHANNEL_LIMIT {
-            tx.try_send(bitcoin_rs_p2p::InboundBlock::from_decoded(block.clone()))
-                .unwrap_or_else(|e| panic!("send within the bound must succeed: {e}"));
-        }
-        let overflow = tx.try_send(bitcoin_rs_p2p::InboundBlock::from_decoded(block));
-        assert!(
-            matches!(overflow, Err(crossbeam_channel::TrySendError::Full(_))),
-            "channel must reject blocks past INBOUND_BLOCK_CHANNEL_LIMIT, got {overflow:?}",
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn inbound_tx_channel_is_bounded_against_flood() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-        let sender = state.inbound_tx_sender();
-        let (lease_tx, _lease_rx) = crossbeam_channel::unbounded();
-        let lease = bitcoin_rs_p2p::PeerLease::new(lease_tx);
-        let source = lease.source(std::net::SocketAddr::from(([127, 0, 0, 1], 8333)));
-        let tx = bitcoin_rs_primitives::Tx {
-            version: 1,
-            inputs: Vec::new(),
-            outputs: Vec::new(),
-            lock_time: 0,
-        };
-        for _ in 0..super::INBOUND_TX_CHANNEL_LIMIT {
-            sender
-                .try_send(bitcoin_rs_p2p::InboundTx::new(tx.clone(), source))
-                .unwrap_or_else(|error| panic!("send within the bound must succeed: {error}"));
-        }
-        let overflow = sender.try_send(bitcoin_rs_p2p::InboundTx::new(tx, source));
-        assert!(
-            matches!(overflow, Err(crossbeam_channel::TrySendError::Full(_))),
-            "channel must reject txs past INBOUND_TX_CHANNEL_LIMIT, got {overflow:?}",
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn open_constructs_full_rpc_handle_set() -> anyhow::Result<()> {
-        use tempfile::tempdir;
-
-        let dir = tempdir()?;
-        let config = crate::NodeConfig {
-            data_dir: dir.path().join("node"),
-            ..crate::NodeConfig::default_for_network(crate::Network::Regtest)
-        };
-
-        let state = NodeState::open(config, None)?;
-        let chain_tip = state.chain_tip();
-        let blocks = state.blocks();
-        let transactions = state.transactions();
-        let network = state.network();
-
-        assert!(chain_tip.load().is_none(), "fresh chain tip must be empty");
-        assert!(blocks.read().is_empty(), "fresh blocks must be empty");
-        assert!(
-            transactions.read().is_empty(),
-            "fresh transactions must be empty"
-        );
-        assert_eq!(network.read().connection_count, 0);
-
-        Ok(())
-    }
-
-    #[test]
-    fn apply_handles_follow_txindex_availability() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("without-txindex");
-        config.p2p.listen.clear();
-        config.indexes.txindex = false;
-        let state = NodeState::open(config, None)?;
-        assert!(state.chain_followers().effects().tx_index().is_none());
-
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("with-txindex");
-        config.p2p.listen.clear();
-        config.indexes.txindex = true;
-        let state = NodeState::open(config, None)?;
-        assert!(state.chain_followers().effects().tx_index().is_some());
-        Ok(())
-    }
-
-    #[test]
-    fn prune_service_is_absent_when_config_disables_pruning() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.storage.prune_target_mb = 0;
-
-        let state = NodeState::open(config, None)?;
-
-        assert!(state.prune_service().is_none());
-        Ok(())
-    }
-
-    #[test]
-    fn apply_block_persists_body_under_pruning_key_when_pruning_disabled() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.storage.prune_target_mb = 0;
-        let state = NodeState::open(config, None)?;
-        let block = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        let hash = Hash256::from_le_bytes(block.block_hash().as_bytes());
-
-        assert!(state.prune_service().is_none());
-        state.apply_block(&block)?;
-
-        assert_eq!(
-            state.blocks.read().first().map(|record| record.body_size),
-            Some(consensus_bytes(&block).len())
-        );
-        assert_eq!(
-            state.block_body_store.load_block_body(0, hash)?.as_deref(),
-            Some(consensus_bytes(&block).as_slice())
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn persisting_same_block_body_twice_appends_once() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-        let hash = bitcoin_rs_primitives::Hash256::from_le_bytes(&[7_u8; 32]);
-        let body = b"idempotent block body";
-
-        state.block_body_store.persist_block_body(42, hash, body)?;
-        let block_file = state.data_dir.join("blocks").join("blk00000.dat");
-        let first_len = std::fs::metadata(&block_file)?.len();
-        state.block_body_store.persist_block_body(42, hash, body)?;
-        let second_len = std::fs::metadata(block_file)?.len();
-
-        assert_eq!(second_len, first_len);
-        assert_eq!(
-            state.block_body_store.load_block_body(42, hash)?.as_deref(),
-            Some(body.as_slice())
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn new_datadir_initializes_current_schema_before_storage() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        std::fs::create_dir_all(&config.data_dir)?;
-        std::fs::write(config.data_dir.join(".CURRENT_SCHEMA.tmp"), b"partial")?;
-
-        let data_dir = config.data_dir.clone();
-        let _state = NodeState::open(config, None)?;
-        assert_eq!(
-            std::fs::read(data_dir.join(crate::checkpoint_fs::CURRENT_SCHEMA_FILE))?,
-            crate::checkpoint_fs::current_schema_bytes()
-        );
-        assert!(!data_dir.join(".CURRENT_SCHEMA.tmp").exists());
-        assert!(data_dir.join("chainstate").exists());
-        Ok(())
-    }
-
-    #[test]
-    fn unmarked_nonempty_datadir_adopts_baseline_schema() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("legacy-node");
-        config.p2p.listen.clear();
-        std::fs::create_dir_all(&config.data_dir)?;
-        std::fs::write(config.data_dir.join("legacy-state"), b"old")?;
-
-        let data_dir = config.data_dir.clone();
-        let _state = NodeState::open(config, None)?;
-        assert_eq!(
-            std::fs::read(data_dir.join(crate::checkpoint_fs::CURRENT_SCHEMA_FILE))?,
-            b"0\n"
-        );
-        assert!(
-            data_dir.join("chainstate").exists(),
-            "baseline adoption must initialize storage"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn mismatched_datadir_schema_is_refused_before_storage_opens() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("old-node");
-        config.p2p.listen.clear();
-        std::fs::create_dir_all(&config.data_dir)?;
-        std::fs::write(
-            config
-                .data_dir
-                .join(crate::checkpoint_fs::CURRENT_SCHEMA_FILE),
-            b"1\n",
-        )?;
-
-        let data_dir = config.data_dir.clone();
-        let Err(error) = NodeState::open(config, None) else {
-            anyhow::bail!("mismatched datadir schema unexpectedly opened");
-        };
-        let message = format!("{error:#}");
-        assert!(message.contains("CURRENT_SCHEMA is not the current datadir schema epoch"));
-        assert!(message.contains("full resync"));
-        assert!(!data_dir.join("chainstate").exists());
-        Ok(())
-    }
-
-    #[test]
-    fn apply_block_with_serialized_persists_same_body_as_apply_block() -> anyhow::Result<()> {
-        let block = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        let hash = Hash256::from_le_bytes(block.block_hash().as_bytes());
-        let serialized = bytes::Bytes::from(consensus_bytes(&block));
-
-        let dir_a = tempfile::tempdir()?;
-        let mut config_a = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config_a.data_dir = dir_a.path().join("node-a");
-        config_a.p2p.listen.clear();
-        config_a.storage.prune_target_mb = 0;
-        let state_a = NodeState::open(config_a, None)?;
-        state_a.apply_block(&block)?;
-
-        let dir_b = tempfile::tempdir()?;
-        let mut config_b = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config_b.data_dir = dir_b.path().join("node-b");
-        config_b.p2p.listen.clear();
-        config_b.storage.prune_target_mb = 0;
-        let state_b = NodeState::open(config_b, None)?;
-        state_b
-            .chainstate()
-            .apply_block_with_serialized(&block, serialized)?;
-
-        let body_a = state_a
-            .block_body_store
-            .load_block_body(0, hash)?
-            .ok_or_else(|| anyhow::anyhow!("apply_block body missing"))?;
-        let body_b = state_b
-            .block_body_store
-            .load_block_body(0, hash)?
-            .ok_or_else(|| anyhow::anyhow!("apply_block_with_serialized body missing"))?;
-        assert_eq!(body_a, body_b);
-        Ok(())
-    }
-
-    /// Undo pruning must respect the durable tip, not the in-memory one.
-    ///
-    /// The applied tip can run far ahead of the last clean checkpoint. Pruning
-    /// undo to the in-memory tip deletes the record for the block the
-    /// checkpoint names, and a crash then restores a chainstate that cannot
-    /// disconnect its own tip.
-    #[test]
-    fn the_chain_transaction_count_survives_a_checkpoint_restart() -> anyhow::Result<()> {
-        use std::sync::atomic::Ordering;
-
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-
-        let expected = {
-            let state = NodeState::open(config.clone(), None)?;
-            assert_eq!(
-                state.chain_tx_count_handle().load(Ordering::Relaxed),
-                0,
-                "a node that has applied nothing cannot know the count"
-            );
-
-            let genesis = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-            let genesis_tx_count = u64::try_from(genesis.txs.len())?;
-            let _tip = state.apply_block(&genesis)?;
-            let counted = state.chain_tx_count_handle().load(Ordering::Relaxed);
-            assert_eq!(counted, genesis_tx_count, "genesis establishes the count");
-
-            assert!(matches!(
-                state.write_clean_checkpoint()?,
-                crate::checkpoint::CheckpointWrite::Published { .. }
-            ));
-            counted
-        };
-
-        // The block-record log is rebuilt empty on every open, so before this
-        // change the count was unrecoverable after a restart: the applied tip
-        // came back from the checkpoint at its real height with nothing behind
-        // it to sum. The number now rides along with the tip.
-        let resumed = NodeState::open(config, None)?;
-        assert_eq!(resumed.resume_source(), ResumeSource::Checkpoint);
-        assert!(
-            resumed.blocks().read().is_empty(),
-            "the record log really does start empty; the count cannot come from it"
-        );
-        assert_eq!(
-            resumed.chain_tx_count_handle().load(Ordering::Relaxed),
-            expected
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn undo_pruning_keeps_records_the_durable_tip_still_needs() -> anyhow::Result<()> {
-        fn hash(height: u32) -> anyhow::Result<bitcoin_rs_primitives::Hash256> {
-            let byte = u8::try_from(height)
-                .map_err(|_| anyhow::anyhow!("test height {height} exceeds u8"))?;
-            Ok(bitcoin_rs_primitives::Hash256::from_le_bytes(&[byte; 32]))
-        }
-
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.storage.prune_target_mb = 1;
-        let state = NodeState::open(config, None)?;
-        publish_applied_tip_height(&state, 11 + CORE_REORG_SAFETY_MARGIN);
-
-        for height in 10_u32..=12 {
-            let hash = hash(height)?;
-            state
-                .block_body_store
-                .persist_block_body(height, hash, b"block-body")?;
-            state
-                .chainstate()
-                .undo_store
-                .persist_undo(height, hash, b"undo-body")?;
-            state.blocks.write().push(BlockRecord {
-                hash: BlockHash::from(hash),
-                height,
-                body_size: 1,
-                header: None,
-                tx_count: 0,
-                time: 0,
-            });
-        }
-
-        // No checkpoint has been written, so nothing is durable above genesis.
-        assert_eq!(
-            state.durable_tip_height.load(Ordering::Acquire),
-            0,
-            "the fixture must have no durable checkpoint, or this proves nothing"
-        );
-
-        let Some(service) = state.prune_service() else {
-            anyhow::bail!("prune service should exist when prune_target_mb > 0");
-        };
-        let result = service
-            .prune_to_height(11)
-            .map_err(|err| anyhow::anyhow!("prune failed: {err}"))?;
-
-        assert_eq!(
-            result.undo_rows_removed, 0,
-            "no undo record may go while a crash would restore below all of them"
-        );
-        assert!(
-            state.storage.stored_prune_undo(10, hash(10)?)?.is_some(),
-            "the record a restore would need must survive"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn prune_service_deletes_seeded_storage_rows_and_advances_pruneheight() -> anyhow::Result<()> {
-        fn hash(height: u32) -> anyhow::Result<bitcoin_rs_primitives::Hash256> {
-            let byte = u8::try_from(height)
-                .map_err(|_| anyhow::anyhow!("test height {height} exceeds u8"))?;
-            Ok(bitcoin_rs_primitives::Hash256::from_le_bytes(&[byte; 32]))
-        }
-
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.storage.prune_target_mb = 1;
-        let state = NodeState::open(config, None)?;
-        publish_applied_tip_height(&state, 11 + CORE_REORG_SAFETY_MARGIN);
-
-        for height in 10_u32..=12 {
-            let hash = hash(height)?;
-            state
-                .block_body_store
-                .persist_block_body(height, hash, b"block-body")?;
-            state
-                .chainstate()
-                .undo_store
-                .persist_undo(height, hash, b"undo-body")?;
-            state.blocks.write().push(BlockRecord {
-                hash: BlockHash::from(hash),
-                height,
-                body_size: 1,
-                header: None,
-                tx_count: 0,
-                time: 0,
-            });
-        }
-
-        // A node pruning old history has a durable tip far above it. Undo
-        // records within the reorg-safety margin of that tip are kept, so the
-        // durable tip has to clear height 11 by more than the margin for this
-        // prune to touch anything. Without that the prune is correctly refused,
-        // which the sibling test asserts.
-        state
-            .durable_tip_height
-            .store(11 + CORE_REORG_SAFETY_MARGIN, Ordering::Release);
-
-        let Some(service) = state.prune_service() else {
-            anyhow::bail!("prune service should exist when prune_target_mb > 0");
-        };
-        let result = service
-            .prune_to_height(11)
-            .map_err(|err| anyhow::anyhow!("prune failed: {err}"))?;
-
-        assert_eq!(result.pruneheight, 11);
-        assert_eq!(result.block_rows_removed, 1);
-        assert_eq!(result.undo_rows_removed, 1);
-        assert!(state.storage.stored_prune_body(10, hash(10)?)?.is_none());
-        assert!(state.storage.stored_prune_undo(10, hash(10)?)?.is_none());
-        assert!(state.storage.stored_prune_body(11, hash(11)?)?.is_some());
-        assert!(state.storage.stored_prune_undo(11, hash(11)?)?.is_some());
-        assert!(state.storage.stored_prune_body(12, hash(12)?)?.is_some());
-        assert!(state.storage.stored_prune_undo(12, hash(12)?)?.is_some());
-
-        Ok(())
-    }
-
-    #[test]
-    fn prune_reclaims_whole_files_and_keeps_current_file() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.storage.prune_target_mb = 1;
-        std::fs::create_dir_all(&config.data_dir)?;
-        let blocks_dir = config.data_dir.join("blocks");
-        std::fs::create_dir_all(&blocks_dir)?;
-        let prunable_file = blocks_dir.join("blk00000.dat");
-        let current_file = blocks_dir.join("blk00001.dat");
-        std::fs::write(&prunable_file, [])?;
-        std::fs::write(&current_file, [])?;
-        let state = NodeState::open(config, None)?;
-        publish_applied_tip_height(&state, 11 + CORE_REORG_SAFETY_MARGIN);
-        // The prune is a no-op until a durable tip is published;
-        // `bitcoin_rs_storage::pruning::stage_block_and_undo_prune` owns file selection.
-        state
-            .durable_tip_height
-            .store(11 + CORE_REORG_SAFETY_MARGIN, Ordering::Release);
-        let hash = bitcoin_rs_primitives::Hash256::from_le_bytes(&[10_u8; 32]);
-        let position = bitcoin_rs_storage::BlockFilePosition {
-            file_no: 0,
-            offset: 0,
-            len: 0,
-        };
-        state.storage.write_test_rows(&[
-            (
-                bitcoin_rs_storage::pruning::BLOCK_DATA_CF,
-                bitcoin_rs_storage::pruning::block_body_key(10, hash).to_vec(),
-                position.encode().to_vec(),
-            ),
-            (
-                bitcoin_rs_storage::pruning::BLOCK_DATA_CF,
-                bitcoin_rs_storage::block_file_max_height_key(0).to_vec(),
-                bitcoin_rs_storage::encode_block_file_max_height(10).to_vec(),
-            ),
-        ])?;
-        let Some(service) = state.prune_service() else {
-            anyhow::bail!("prune service should exist when prune_target_mb > 0");
-        };
-        service
-            .prune_to_height(11)
-            .map_err(|error| anyhow::anyhow!("prune failed: {error}"))?;
-
-        assert!(!prunable_file.exists());
-        assert!(current_file.exists());
-        assert!(state.storage.stored_prune_body(10, hash)?.is_none());
-        let has_metadata = state
-            .storage
-            .read_test_row(
-                bitcoin_rs_storage::pruning::BLOCK_DATA_CF,
-                &bitcoin_rs_storage::block_file_max_height_key(0),
-            )?
-            .is_some();
-        assert!(!has_metadata);
-        Ok(())
-    }
-
-    /// Pruning a block file must reduce what the node reports as its disk size.
-    ///
-    /// `getblockchaininfo.size_on_disk` used to be the sum of every block
-    /// record's `body_size`. Pruning does not remove records — it clears their
-    /// cached bodies and leaves the rest — so that sum could not move, and a
-    /// pruned node went on reporting bytes it no longer had, under the one field
-    /// an operator reads to check that pruning worked.
-    ///
-    /// This asserts both halves: the store's figure falls by exactly the file
-    /// that was deleted, and the record sum does not move at all. The second is
-    /// what makes the first worth having.
-    #[test]
-    fn pruning_a_block_file_reduces_the_reported_disk_size() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.storage.prune_target_mb = 1;
-
-        std::fs::create_dir_all(&config.data_dir)?;
-
-        // Two files present before the store opens, so the earlier one is not
-        // the append target and is therefore prunable. Same shape as
-        // `prune_reclaims_whole_files_and_keeps_current_file`, but with bytes in
-        // it, because bytes are what is being counted.
-        let blocks_dir = config.data_dir.join("blocks");
-        std::fs::create_dir_all(&blocks_dir)?;
-        let prunable_file = blocks_dir.join("blk00000.dat");
-        let prunable_bytes = vec![7_u8; 4_096];
-        std::fs::write(&prunable_file, &prunable_bytes)?;
-        std::fs::write(blocks_dir.join("blk00001.dat"), [])?;
-
-        let state = NodeState::open(config, None)?;
-        publish_applied_tip_height(&state, 11 + CORE_REORG_SAFETY_MARGIN);
-        // The prune is a no-op until a durable tip is published;
-        // `bitcoin_rs_storage::pruning::stage_block_and_undo_prune` owns file selection.
-        state
-            .durable_tip_height
-            .store(11 + CORE_REORG_SAFETY_MARGIN, Ordering::Release);
-        let block = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        // The hash is not needed: this test counts bytes in files, not bodies.
-        let record = BlockRecord::from_block(10, &block);
-        let record_sum_before = u64::try_from(record.body_size)?;
-        state.blocks.write().push(record);
-
-        let Some(before) = state.block_body_store.disk_usage() else {
-            anyhow::bail!("a flat-file store must report its usage");
-        };
-        assert!(
-            before >= u64::try_from(prunable_bytes.len())?,
-            "the fixture's bytes must be accounted for"
-        );
-
-        // Tell the pruner that file 0 tops out at height 10, so pruning to 11
-        // makes it prunable.
-        state.storage.write_test_rows(&[(
-            bitcoin_rs_storage::pruning::BLOCK_DATA_CF,
-            bitcoin_rs_storage::block_file_max_height_key(0).to_vec(),
-            bitcoin_rs_storage::encode_block_file_max_height(10).to_vec(),
-        )])?;
-
-        let Some(service) = state.prune_service() else {
-            anyhow::bail!("prune service should exist when prune_target_mb > 0");
-        };
-        service
-            .prune_to_height(11)
-            .map_err(|error| anyhow::anyhow!("prune failed: {error}"))?;
-
-        assert!(!prunable_file.exists(), "the fixture must actually prune");
-        let Some(after) = state.block_body_store.disk_usage() else {
-            anyhow::bail!("a flat-file store must report its usage");
-        };
-        assert_eq!(
-            after,
-            before.saturating_sub(u64::try_from(prunable_bytes.len())?),
-            "the reported size must fall by exactly the file that was deleted"
-        );
-
-        // The number this replaces, unmoved — which is the defect.
-        let record_sum_after = state.blocks.read().iter().fold(0_u64, |total, entry| {
-            total.saturating_add(u64::try_from(entry.body_size).unwrap_or(0))
-        });
-        assert_eq!(
-            record_sum_after, record_sum_before,
-            "the block-record sum cannot see pruning, which is why it is not \
-             what size_on_disk reports"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn manual_prune_removes_pruned_block_transactions_from_cache() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.storage.prune_target_mb = 1;
-        let state = NodeState::open(config, None)?;
-        publish_applied_tip_height(&state, 11 + CORE_REORG_SAFETY_MARGIN);
-
-        let pruned_block = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        let pruned_hash = Hash256::from_le_bytes(pruned_block.block_hash().as_bytes());
-        state.block_body_store.persist_block_body(
-            10,
-            pruned_hash,
-            &consensus_bytes(&pruned_block),
-        )?;
-        state
-            .chainstate()
-            .undo_store
-            .persist_undo(10, pruned_hash, b"undo-body")?;
-        state
-            .blocks
-            .write()
-            .push(BlockRecord::from_block(10, &pruned_block));
-
-        let pruned_tx = pruned_block.txs[0].clone();
-        let pruned_txid = pruned_tx.txid();
-        let unrelated_tx = Tx {
-            version: 2,
-            lock_time: 0,
-            inputs: Vec::new(),
-            outputs: Vec::new(),
-        };
-        let unrelated_txid = unrelated_tx.txid();
-
-        {
-            let mut transactions = state.transactions.write();
-            transactions.insert(pruned_txid, pruned_tx);
-            transactions.insert(unrelated_txid, unrelated_tx);
-        }
-
-        let Some(service) = state.prune_service() else {
-            anyhow::bail!("prune service should exist when prune_target_mb > 0");
-        };
-        service
-            .prune_to_height(11)
-            .map_err(|err| anyhow::anyhow!("prune failed: {err}"))?;
-
-        let transactions = state.transactions.read();
-        assert!(!transactions.contains_key(&pruned_txid));
-        assert!(transactions.contains_key(&unrelated_txid));
-        Ok(())
-    }
-
-    #[test]
-    fn prune_service_restores_persisted_pruneheight_on_reopen() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.storage.prune_target_mb = 1;
-
-        {
-            let state = NodeState::open(config.clone(), None)?;
-            publish_applied_tip_height(&state, 11 + CORE_REORG_SAFETY_MARGIN);
-            let Some(service) = state.prune_service() else {
-                anyhow::bail!("prune service should exist when prune_target_mb > 0");
-            };
-            let result = service
-                .prune_to_height(11)
-                .map_err(|err| anyhow::anyhow!("prune failed: {err}"))?;
-            assert_eq!(result.pruneheight, 11);
-        }
-
-        let reopened = NodeState::open(config, None)?;
-        let Some(service) = reopened.prune_service() else {
-            anyhow::bail!("prune service should exist when prune_target_mb > 0");
-        };
-        assert_eq!(service.status().pruneheight, Some(11));
-
-        Ok(())
-    }
-
-    #[test]
-    fn prune_waits_for_chain_transition_and_revalidates_applied_tip() -> anyhow::Result<()> {
-        use std::time::Duration;
-
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.storage.prune_target_mb = 1;
-        let state = NodeState::open(config.clone(), None)?;
-        publish_applied_tip_height(&state, 11 + CORE_REORG_SAFETY_MARGIN);
-        let Some(service) = state.prune_service() else {
-            anyhow::bail!("prune service should exist when prune_target_mb > 0");
-        };
-
-        let handles = state.chainstate();
-        let transition = handles.chain_transition.lock();
-        let (started_tx, started_rx) = std::sync::mpsc::sync_channel(1);
-        let (done_tx, done_rx) = std::sync::mpsc::sync_channel(1);
-        let (done_while_locked, result) = std::thread::scope(|scope| -> anyhow::Result<_> {
-            let service = Arc::clone(&service);
-            let worker = scope.spawn(move || {
-                let _ = started_tx.send(());
-                let result = service.prune_to_height(11);
-                let _ = done_tx.send(());
-                result
-            });
-            started_rx.recv_timeout(Duration::from_secs(5))?;
-            let done_while_locked = done_rx.recv_timeout(Duration::from_millis(100));
-            publish_applied_tip_height(&state, 10 + CORE_REORG_SAFETY_MARGIN);
-            drop(transition);
-            let result = worker
-                .join()
-                .map_err(|_| anyhow::anyhow!("prune worker panicked"))?;
-            Ok((done_while_locked, result))
-        })?;
-        let status_after = service.status();
-        drop(service);
-        drop(handles);
-        drop(state);
-        let reopened = NodeState::open(config, None)?;
-        let Some(reopened_service) = reopened.prune_service() else {
-            anyhow::bail!("prune service should exist after reopen");
-        };
-
-        assert!(
-            matches!(
-                done_while_locked,
-                Err(std::sync::mpsc::RecvTimeoutError::Timeout)
-            ),
-            "pruning entered while another chain transition held authority"
-        );
-        assert!(
-            matches!(
-                &result,
-                Err(PruneServiceError::Failed(message))
-                    if message == "prune height is within reorg safety margin"
-            ),
-            "pruning did not reject the tip observed under authority: {result:?}"
-        );
-        assert_eq!(status_after.pruneheight, None);
-        assert_eq!(reopened_service.status().pruneheight, None);
-        Ok(())
-    }
-
-    #[test]
-    fn prune_refuses_after_apply_admission_closes() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        config.storage.prune_target_mb = 1;
-        let state = NodeState::open(config, None)?;
-        publish_applied_tip_height(&state, 11 + CORE_REORG_SAFETY_MARGIN);
-        let Some(service) = state.prune_service() else {
-            anyhow::bail!("prune service should exist when prune_target_mb > 0");
-        };
-
-        state.apply_handles.admission.close_permanently();
-        let result = service.prune_to_height(11);
-
-        assert!(
-            matches!(
-                &result,
-                Err(PruneServiceError::Failed(message))
-                    if message == "block apply rejected because clean shutdown has begun"
-            ),
-            "pruning ignored closed chain-mutation admission: {result:?}"
-        );
-        assert_eq!(service.status().pruneheight, None);
-        Ok(())
-    }
-
-    /// Overlapping prune calls must commit in acquisition order.
-    ///
-    /// A lower request blocks inside body-store load while holding
-    /// `pruneheight`; a higher contender must neither enter that load nor
-    /// finish until the lower call releases, and both persisted and in-memory
-    /// pruneheight end at the higher value.
-    #[cfg(feature = "fjall")]
-    #[allow(clippy::too_many_lines)]
-    #[test]
-    fn prune_to_height_serializes_overlapping_calls() -> anyhow::Result<()> {
-        use std::sync::Barrier;
-        use std::sync::atomic::AtomicUsize;
-        use std::time::Duration;
-
-        struct BlockingPruneBodyStore {
-            entered: Barrier,
-            release: Barrier,
-            block_once: AtomicBool,
-            loads: AtomicUsize,
-        }
-
-        impl bitcoin_rs_storage::block_body::BlockBodyStore for BlockingPruneBodyStore {
-            fn load_block_body(
-                &self,
-                _height: u32,
-                _hash: bitcoin_rs_primitives::Hash256,
-            ) -> Result<Option<Vec<u8>>, bitcoin_rs_storage::StorageError> {
-                self.loads.fetch_add(1, Ordering::AcqRel);
-                if self.block_once.swap(false, Ordering::AcqRel) {
-                    self.entered.wait();
-                    self.release.wait();
-                }
-                Ok(None)
-            }
-
-            fn persist_block_body(
-                &self,
-                _height: u32,
-                _hash: bitcoin_rs_primitives::Hash256,
-                _body: &[u8],
-            ) -> Result<(), bitcoin_rs_storage::StorageError> {
-                Ok(())
-            }
-
-            fn sync(&self) -> Result<(), bitcoin_rs_storage::StorageError> {
-                Ok(())
-            }
-        }
-
-        let dir = tempfile::tempdir()?;
-        let mut authority_config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        authority_config.data_dir = dir.path().join("authority");
-        authority_config.p2p.listen.clear();
-        let authority_state = NodeState::open(authority_config, None)?;
-        publish_applied_tip_height(&authority_state, 12 + CORE_REORG_SAFETY_MARGIN);
-        let data_dir = dir.path().join("node");
-        std::fs::create_dir_all(data_dir.join("chainstate"))?;
-        let store = Arc::new(bitcoin_rs_storage::FjallStore::open(
-            data_dir.join("chainstate"),
-        )?);
-        let block_files = Arc::new(FlatFileBlockStore::open(&data_dir)?);
-        let body_store = Arc::new(BlockingPruneBodyStore {
-            entered: Barrier::new(2),
-            release: Barrier::new(2),
-            block_once: AtomicBool::new(true),
-            loads: AtomicUsize::new(0),
-        });
-        let body_handle: Arc<dyn bitcoin_rs_storage::block_body::BlockBodyStore> =
-            body_store.clone();
-        let blocks = Arc::new(RwLock::new(BlockLog::new()));
-        let hash = bitcoin_rs_primitives::Hash256::from_le_bytes(&[10_u8; 32]);
-        blocks.write().push(BlockRecord {
-            hash: BlockHash::from(hash),
-            height: 10,
-            body_size: 1,
-            header: None,
-            tx_count: 1,
-            time: 0,
-        });
-        let service = Arc::new(NodePruneService::new(
-            Arc::clone(&store),
-            block_files,
-            body_handle,
-            Arc::clone(&blocks),
-            Arc::new(RwLock::new(HashMap::new())),
-            authority_state.chainstate().prune_authority(),
-            Arc::new(AtomicU32::new(11 + CORE_REORG_SAFETY_MARGIN)),
-        )?);
-
-        let (started_tx, started_rx) = std::sync::mpsc::sync_channel(1);
-        let (done_tx, done_rx) = std::sync::mpsc::sync_channel(1);
-        std::thread::scope(|scope| -> anyhow::Result<()> {
-            let lower_service = Arc::clone(&service);
-            let lower = scope.spawn(move || lower_service.prune_to_height(11));
-            body_store.entered.wait();
-
-            let higher_service = Arc::clone(&service);
-            let higher = scope.spawn(move || {
-                let _ = started_tx.send(());
-                let result = higher_service.prune_to_height(12);
-                let _ = done_tx.send(());
-                result
-            });
-            let higher_started = started_rx.recv_timeout(Duration::from_secs(5));
-            let higher_done = done_rx.recv_timeout(Duration::from_millis(100));
-            let loads_while_lower_blocked = body_store.loads.load(Ordering::Acquire);
-            body_store.release.wait();
-            let lower_join = lower.join();
-            let higher_join = higher.join();
-            higher_started
-                .map_err(|error| anyhow::anyhow!("higher prune did not start: {error}"))?;
-            let lower_result = lower_join
-                .map_err(|_| anyhow::anyhow!("lower prune panicked"))?
-                .map_err(|err| anyhow::anyhow!("lower prune failed: {err}"))?;
-            let higher_result = higher_join
-                .map_err(|_| anyhow::anyhow!("higher prune panicked"))?
-                .map_err(|err| anyhow::anyhow!("higher prune failed: {err}"))?;
-            assert!(
-                matches!(higher_done, Err(std::sync::mpsc::RecvTimeoutError::Timeout)),
-                "higher prune committed while lower still held the pruneheight lock; done_rx={higher_done:?}"
-            );
-            assert_eq!(
-                loads_while_lower_blocked, 1,
-                "higher prune must not enter body-store load while lower holds pruneheight; loads={loads_while_lower_blocked}"
-            );
-            assert_eq!(lower_result.pruneheight, 11);
-            assert_eq!(higher_result.pruneheight, 12);
-            Ok(())
-        })?;
-
-        assert_eq!(service.status().pruneheight, Some(12));
-        assert_eq!(load_pruneheight(&*store)?, Some(12));
-        Ok(())
-    }
-
-    #[test]
-    fn clean_checkpoint_reopens_and_applies_the_next_block() -> anyhow::Result<()> {
-        fn stable_hash(
-            view: &bitcoin_rs_utxo::UtxoSetView<'_>,
-        ) -> Result<bitcoin_rs_primitives::Hash256, bitcoin_rs_utxo::UtxoError> {
-            view.hash_serialized_3()
-        }
-        let dir = tempfile::tempdir()?;
-        let data_dir = dir.path().join("node");
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = data_dir.clone();
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-        let genesis = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        let genesis_tip = state.apply_block(&genesis)?;
-        let expected_utxo_hash = state.utxo().with_stable_view(stable_hash)?;
-        let expected_stats = state.coin_stats().snapshot();
-        assert!(matches!(
-            state.write_clean_checkpoint()?,
-            crate::checkpoint::CheckpointWrite::Published { .. }
-        ));
-        drop(state);
-
-        let mut reopen_config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        reopen_config.data_dir = data_dir.clone();
-        reopen_config.p2p.listen.clear();
-        let resumed = NodeState::open(reopen_config.clone(), None)?;
-        assert_eq!(resumed.resume_source(), ResumeSource::Checkpoint);
-        let applied = resumed
-            .applied_tip()
-            .load_full()
-            .ok_or_else(|| std::io::Error::other("checkpoint did not publish applied tip"))?;
-        assert_eq!(applied.height, genesis_tip.height);
-        assert_eq!(applied.hash, genesis_tip.hash);
-        assert_eq!(
-            resumed.chain_tip().load_full().as_deref(),
-            Some(applied.as_ref())
-        );
-        assert_eq!(
-            resumed.utxo().with_stable_view(stable_hash)?,
-            expected_utxo_hash
-        );
-        assert_eq!(resumed.coin_stats().snapshot(), expected_stats);
-        assert!(resumed.blocks().read().is_empty());
-        assert!(resumed.transactions().read().is_empty());
-        assert!(resumed.mempool().read().is_empty());
-
-        let next = mined_regtest_child(genesis.block_hash())?;
-        let next_tip = resumed.apply_block(&next)?;
-        assert_eq!(next_tip.height, 1);
-        assert_eq!(
-            next_tip.hash.to_le_bytes(),
-            next.block_hash().0.to_le_bytes()
-        );
-        let listener_after_apply = resumed.coin_stats().snapshot();
-        let mut rescanned = resumed.utxo().with_stable_view(|view| {
-            bitcoin_rs_utxo::stats::scan_coin_stats(view, next_tip.height, true)
-        })?;
-        rescanned.tx_count = listener_after_apply.tx_count;
-        assert_eq!(
-            listener_after_apply.total_amount, rescanned.total_amount,
-            "checkpoint resume must keep rolling CoinStats attached to UTXO commits"
-        );
-        resumed.write_clean_checkpoint()?;
-
-        let root = data_dir.join("chainstate-checkpoints");
-        let current: serde_json::Value =
-            serde_json::from_slice(&std::fs::read(root.join("CURRENT"))?)?;
-        let directory = current
-            .get("directory")
-            .and_then(serde_json::Value::as_str)
-            .ok_or_else(|| std::io::Error::other("CURRENT has no generation directory"))?;
-        let snapshot_file = std::fs::File::open(root.join(directory).join("utxo-v4.dat"))?;
-        let mut snapshot_reader = std::io::BufReader::new(snapshot_file);
-        let snapshot = bitcoin_rs_utxo::snapshot::read_snapshot_strict_v4(&mut snapshot_reader)?;
-        assert_ne!(snapshot.muhash_trailer, [0_u8; 384]);
-        assert_eq!(snapshot.muhash_trailer, rescanned.muhash.finalize());
-        drop(resumed);
-
-        let resumed_again = NodeState::open(reopen_config, None)?;
-        assert_eq!(resumed_again.resume_source(), ResumeSource::Checkpoint);
-        assert_eq!(resumed_again.coin_stats().snapshot(), rescanned);
-        Ok(())
-    }
-
-    #[test]
-    fn clean_checkpoint_lifecycle_is_backend_neutral() -> anyhow::Result<()> {
-        let backends = vec![
-            #[cfg(feature = "fjall")]
-            "fjall",
-            #[cfg(feature = "rocksdb")]
-            "rocksdb",
-            #[cfg(feature = "redb")]
-            "redb",
-        ];
-
-        for backend in backends {
-            let dir = tempfile::tempdir()?;
-            let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-            config.data_dir = dir.path().join(backend);
-            config.storage.backend = backend.parse().map_err(anyhow::Error::msg)?;
-            config.p2p.listen.clear();
-            let genesis = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-            let state = NodeState::open(config.clone(), None)?;
-            state.apply_block(&genesis)?;
-            state.write_clean_checkpoint()?;
-            drop(state);
-
-            let resumed = NodeState::open(config, None)?;
-            assert_eq!(resumed.resume_source(), ResumeSource::Checkpoint);
-            resumed.apply_block(&mined_regtest_child(genesis.block_hash())?)?;
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn rolling_coinstats_resume_continues_through_next_block() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let data_dir = dir.path().join("node-g2");
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = data_dir.clone();
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-        let genesis = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        state.apply_block(&genesis)?;
-        let before = state.coin_stats().snapshot();
-        state.write_clean_checkpoint()?;
-        drop(state);
-
-        let mut reopen_config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        reopen_config.data_dir = data_dir;
-        reopen_config.p2p.listen.clear();
-        let resumed = NodeState::open(reopen_config, None)?;
-        assert_eq!(resumed.coin_stats().snapshot(), before);
-        resumed.apply_block(&mined_regtest_child(genesis.block_hash())?)?;
-        let rolling = resumed.coin_stats().snapshot();
-        let mut scanned = resumed.utxo().with_stable_view(|view| {
-            bitcoin_rs_utxo::stats::scan_coin_stats(view, rolling.height, true)
-        })?;
-        scanned.tx_count = rolling.tx_count;
-        // The listener is attached only after checkpoint/journal restoration,
-        // then tracks subsequent live UTXO mutations without double-counting
-        // the restored snapshot.
-        assert_eq!(
-            rolling.total_amount, scanned.total_amount,
-            "restored state must receive subsequent rolling UTXO notifications"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn journal_replay_restores_state_above_checkpoint() -> anyhow::Result<()> {
-        fn stable_hash(
-            view: &bitcoin_rs_utxo::UtxoSetView<'_>,
-        ) -> Result<bitcoin_rs_primitives::Hash256, bitcoin_rs_utxo::UtxoError> {
-            view.hash_serialized_3()
-        }
-
-        let dir = tempfile::tempdir()?;
-        let data_dir = dir.path().join("journal-resume");
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = data_dir;
-        config.p2p.listen.clear();
-        config.chainstate_journal.blocks = 1;
-
-        let genesis = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        let first = NodeState::open(config.clone(), None)?;
-        first.apply_block(&genesis)?;
-        first.write_clean_checkpoint()?;
-        drop(first);
-
-        // The first reopen discards the pre-checkpoint journal generation and
-        // initializes one whose authenticated base is the published checkpoint.
-        let base = NodeState::open(config.clone(), None)?;
-        assert_eq!(base.resume_source(), ResumeSource::Checkpoint);
-        let child = mined_regtest_child(genesis.block_hash())?;
-        let expected_tip = base.apply_block(&child)?;
-        let expected_utxo = base.utxo().with_stable_view(stable_hash)?;
-        let expected_stats = base.coin_stats().snapshot();
-        let expected_tx_count = base.chain_tx_count_handle().load(Ordering::Relaxed);
-        drop(base);
-
-        let resumed = NodeState::open(config, None)?;
-        assert_eq!(resumed.resume_source(), ResumeSource::Journal);
-        let resumed_tip = resumed
-            .applied_tip()
-            .load_full()
-            .ok_or_else(|| std::io::Error::other("journal replay did not publish a tip"))?;
-        assert_eq!(resumed_tip.as_ref(), &expected_tip);
-        assert_eq!(resumed.utxo().with_stable_view(stable_hash)?, expected_utxo);
-        assert_eq!(resumed.coin_stats().snapshot(), expected_stats);
-        assert_eq!(
-            resumed.chain_tx_count_handle().load(Ordering::Relaxed),
-            expected_tx_count
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn shutdown_arc_is_shared_with_apply_handles() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-        assert!(Arc::ptr_eq(&state.shutdown(), &state.chainstate().shutdown));
-        Ok(())
-    }
-
-    #[test]
-    fn checkpoint_refuses_inflight_disconnect_and_preserves_state() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let data_dir = dir.path().join("node");
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = data_dir.clone();
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-        let genesis = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        state.apply_block(&genesis)?;
-        assert!(matches!(
-            state.write_clean_checkpoint()?,
-            crate::checkpoint::CheckpointWrite::Published { .. }
-        ));
-
-        let checkpoint_root = data_dir.join("chainstate-checkpoints");
-        let armed_hash = bitcoin_rs_primitives::Hash256::from_le_bytes(&[0xab; 32]);
-        let armed_height = 10;
-        state
-            .chainstate()
-            .undo_store
-            .arm_disconnect(armed_height, armed_hash)?;
-        let marker_before = state.chainstate().undo_store.load_disconnect_marker()?;
-        let current_before = std::fs::read(checkpoint_root.join("CURRENT"))?;
-        let mut dirs_before = std::collections::BTreeSet::new();
-        for entry in std::fs::read_dir(&checkpoint_root)? {
-            let entry = entry?;
-            if entry.file_type()?.is_dir() {
-                dirs_before.insert(entry.file_name().to_string_lossy().into_owned());
-            }
-        }
-
-        let result = state.write_clean_checkpoint();
-        let Err(crate::checkpoint::CheckpointError::DisconnectInFlight { hash, height }) = result
-        else {
-            anyhow::bail!("expected DisconnectInFlight refusal, got {result:?}");
-        };
-        assert_eq!(hash, armed_hash);
-        assert_eq!(height, armed_height);
-
-        let marker_after = state.chainstate().undo_store.load_disconnect_marker()?;
-        let current_after = std::fs::read(checkpoint_root.join("CURRENT"))?;
-        let mut dirs_after = std::collections::BTreeSet::new();
-        for entry in std::fs::read_dir(&checkpoint_root)? {
-            let entry = entry?;
-            if entry.file_type()?.is_dir() {
-                dirs_after.insert(entry.file_name().to_string_lossy().into_owned());
-            }
-        }
-
-        assert_eq!(marker_before, marker_after);
-        assert_eq!(current_before, current_after);
-        assert_eq!(dirs_before, dirs_after);
-        Ok(())
-    }
-
-    #[test]
-    fn torn_disconnect_refusal_names_authoritative_stores_to_remove() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let data_dir = dir.path().join("node");
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = data_dir.clone();
-        config.p2p.listen.clear();
-        let state = NodeState::open(config.clone(), None)?;
-        state.chainstate().undo_store.arm_disconnect(
-            10,
-            bitcoin_rs_primitives::Hash256::from_le_bytes(&[0xcd; 32]),
-        )?;
-        drop(state);
-
-        let error = match NodeState::open(config, None) {
-            Ok(_) => anyhow::bail!("node reopened with an armed disconnect marker"),
-            Err(error) => error,
-        };
-        let message = error.to_string();
-        for store in ["chainstate", "chainstate-checkpoints", "txindex"] {
-            let path = data_dir.join(store);
-            assert!(
-                message.contains(&path.display().to_string()),
-                "startup refusal omitted {}: {message}",
-                path.display()
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn invalidate_block_settles_disconnect_debt() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let data_dir = dir.path().join("node");
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = data_dir.clone();
-        config.p2p.listen.clear();
-        // Journal rewind disarms disconnect markers itself. These tests own the
-        // checkpoint-settlement path that remains when the journal cannot.
-        config.chainstate_journal.enabled = false;
-        let state = NodeState::open(config, None)?;
-        let genesis = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        state.apply_block(&genesis)?;
-        let block_one = mined_regtest_child_at(genesis.block_hash(), genesis.header.time + 1, 1)?;
-        state.apply_block(&block_one)?;
-        state.publish_checkpoint()?;
-
-        let current_before = serde_json::from_slice::<serde_json::Value>(&std::fs::read(
-            data_dir.join("chainstate-checkpoints/CURRENT"),
-        )?)?
-        .get("generation")
-        .and_then(serde_json::Value::as_u64)
-        .ok_or_else(|| anyhow::anyhow!("CURRENT has no generation"))?;
-        let block_two = mined_regtest_child_at(block_one.block_hash(), genesis.header.time + 2, 2)?;
-        state.apply_block(&block_two)?;
-
-        crate::reorg::invalidate_block(
-            &state.chainstate(),
-            &state.chain_followers(),
-            Hash256::from(block_two.block_hash()),
-        )?;
-
-        assert!(
-            state
-                .chainstate()
-                .undo_store
-                .load_disconnect_marker()?
-                .is_none()
-        );
-        let current_after = serde_json::from_slice::<serde_json::Value>(&std::fs::read(
-            data_dir.join("chainstate-checkpoints/CURRENT"),
-        )?)?
-        .get("generation")
-        .and_then(serde_json::Value::as_u64)
-        .ok_or_else(|| anyhow::anyhow!("CURRENT has no generation"))?;
-        assert!(current_after > current_before);
-        assert_eq!(state.durable_tip_height.load(Ordering::Acquire), 1);
-        Ok(())
-    }
-
-    #[test]
-    fn invalidate_block_settlement_failure_is_not_success() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let data_dir = dir.path().join("node");
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = data_dir;
-        config.p2p.listen.clear();
-        config.chainstate_journal.enabled = false;
-        let state = NodeState::open(config.clone(), None)?;
-        let genesis = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        state.apply_block(&genesis)?;
-        let block_one = mined_regtest_child_at(genesis.block_hash(), genesis.header.time + 1, 1)?;
-        state.apply_block(&block_one)?;
-        state.publish_checkpoint()?;
-
-        let block_two = mined_regtest_child_at(block_one.block_hash(), genesis.header.time + 2, 2)?;
-        state.apply_block(&block_two)?;
-
-        crate::checkpoint::inject_next_checkpoint_failpoint(
-            crate::checkpoint::CheckpointFailpoint::ManifestWrite,
-        );
-        let result = crate::reorg::invalidate_block(
-            &state.chainstate(),
-            &state.chain_followers(),
-            Hash256::from(block_two.block_hash()),
-        );
-        let Err(crate::reorg::ReorgError::CheckpointSettlement(_)) = result else {
-            anyhow::bail!("expected CheckpointSettlement, got {result:?}");
-        };
-
-        let marker = state
-            .chainstate()
-            .undo_store
-            .load_disconnect_marker()?
-            .ok_or_else(|| anyhow::anyhow!("settlement failure cleared the disconnect marker"))?;
-        assert_eq!(
-            marker.phase,
-            bitcoin_rs_storage::DisconnectPhase::RolledBack
-        );
-
-        drop(state);
-        let error = match NodeState::open(config, None) {
-            Ok(_) => anyhow::bail!("node reopened with unsettled RolledBack debt"),
-            Err(error) => error,
-        };
-        assert!(
-            error
-                .to_string()
-                .contains("did not reach a clean checkpoint"),
-            "startup refusal omitted the checkpoint debt: {error}"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn switch_to_branch_settles_disconnect_debt() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let data_dir = dir.path().join("node");
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = data_dir.clone();
-        config.p2p.listen.clear();
-        config.chainstate_journal.enabled = false;
-        let state = NodeState::open(config, None)?;
-        let genesis = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        state.apply_block(&genesis)?;
-        let block_one = mined_regtest_child_at(genesis.block_hash(), genesis.header.time + 1, 1)?;
-        state.apply_block(&block_one)?;
-        state.publish_checkpoint()?;
-
-        let current_before = serde_json::from_slice::<serde_json::Value>(&std::fs::read(
-            data_dir.join("chainstate-checkpoints/CURRENT"),
-        )?)?
-        .get("generation")
-        .and_then(serde_json::Value::as_u64)
-        .ok_or_else(|| anyhow::anyhow!("CURRENT has no generation"))?;
-
-        let genesis_id = state
-            .block_tree
-            .read()
-            .lookup(Hash256::from(genesis.block_hash()))
-            .ok_or_else(|| anyhow::anyhow!("missing genesis node"))?;
-        let mut parent = genesis_id;
-        let mut previous_hash = genesis.block_hash();
-        let mut fork_bodies = HashMap::new();
-        for height in 1..=2 {
-            let block =
-                mined_regtest_child_at(previous_hash, genesis.header.time + 10 + height, height)?;
-            let node_id = state.block_tree.write().insert_node(
-                Some(parent),
-                block.header,
-                bitcoin_rs_chain::node::NodeStatus::HeaderValid,
-            )?;
-            fork_bodies.insert(
-                Hash256::from(block.block_hash()),
-                (block.clone(), bytes::Bytes::from(consensus_bytes(&block))),
-            );
-            parent = node_id;
-            previous_hash = block.block_hash();
-        }
-
-        let handles = state.chainstate();
-        crate::reorg::switch_to_branch(
-            &handles,
-            &state.chain_followers(),
-            parent,
-            |hash| fork_bodies.get(&hash).cloned(),
-            |_| {},
-        )?;
-
-        assert!(
-            state
-                .chainstate()
-                .undo_store
-                .load_disconnect_marker()?
-                .is_none()
-        );
-        let current_after = serde_json::from_slice::<serde_json::Value>(&std::fs::read(
-            data_dir.join("chainstate-checkpoints/CURRENT"),
-        )?)?
-        .get("generation")
-        .and_then(serde_json::Value::as_u64)
-        .ok_or_else(|| anyhow::anyhow!("CURRENT has no generation"))?;
-        assert!(current_after > current_before);
-        assert_eq!(state.durable_tip_height.load(Ordering::Acquire), 2);
-        Ok(())
-    }
-
-    #[test]
-    fn publish_checkpoint_refuses_when_no_applied_tip() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-        let state = NodeState::open(config, None)?;
-        let Err(error) = state.publish_checkpoint() else {
-            anyhow::bail!("checkpoint publication succeeded without an applied tip");
-        };
-        assert!(
-            error.to_string().contains("no applied tip"),
-            "unexpected error: {error}"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn publish_checkpoint_returns_generation_and_reopens() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let data_dir = dir.path().join("node");
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = data_dir;
-        config.p2p.listen.clear();
-        let state = NodeState::open(config.clone(), None)?;
-        let genesis = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        let tip = state.apply_block(&genesis)?;
-        let generation = state.publish_checkpoint()?;
-        assert!(
-            generation > 0,
-            "published checkpoint must have a positive generation"
-        );
-        drop(state);
-
-        let resumed = NodeState::open(config, None)?;
-        assert_eq!(resumed.resume_source(), ResumeSource::Checkpoint);
-        let applied = resumed
-            .applied_tip()
-            .load_full()
-            .ok_or_else(|| std::io::Error::other("checkpoint did not publish applied tip"))?;
-        assert_eq!(applied.height, tip.height);
-        assert_eq!(applied.hash, tip.hash);
-        Ok(())
-    }
-
-    #[test]
-    fn process_epoch_is_strictly_monotonic_across_restart() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-
-        let first = NodeState::open(config.clone(), None)?
-            .active_chain_snapshot()
-            .epoch;
-        let second = NodeState::open(config.clone(), None)?
-            .active_chain_snapshot()
-            .epoch;
-        let third = NodeState::open(config, None)?.active_chain_snapshot().epoch;
-        assert!(first > 0, "a fresh data dir allocates epoch 1, got {first}");
-        assert!(
-            second > first,
-            "restart must never reuse an epoch: {first} -> {second}"
-        );
-        assert!(
-            third > second,
-            "restart must never reuse an epoch: {second} -> {third}"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn active_chain_snapshot_starts_at_genesis_on_fresh_node() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-
-        let state = NodeState::open(config.clone(), None)?;
-        let epoch = state.chain_event_publisher().epoch();
-        assert_eq!(
-            state.active_chain_snapshot(),
-            ChainSnapshot {
-                epoch,
-                sequence: 0,
-                tip_hash: config.network.genesis_block_hash(),
-                tip_height: 0,
-            },
-            "a node that committed nothing anchors at genesis with sequence 0"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn active_chain_snapshot_anchors_at_restored_tip_after_restart() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-
-        let (tip, first_epoch) = {
-            let state = NodeState::open(config.clone(), None)?;
-            let genesis = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-            let tip = state.apply_block(&genesis)?;
-            assert!(matches!(
-                state.write_clean_checkpoint()?,
-                crate::checkpoint::CheckpointWrite::Published { .. }
-            ));
-            (tip, state.active_chain_snapshot().epoch)
-        };
-
-        let resumed = NodeState::open(config, None)?;
-        assert_eq!(resumed.resume_source(), ResumeSource::Checkpoint);
-        let snapshot = resumed.active_chain_snapshot();
-        assert_eq!(snapshot.tip_hash, tip.hash);
-        assert_eq!(snapshot.tip_height, tip.height);
-        assert_eq!(
-            snapshot.sequence, 0,
-            "a restart resets the sequence, never the epoch"
-        );
-        assert!(
-            snapshot.epoch > first_epoch,
-            "restart must advance the epoch: {} -> {}",
-            first_epoch,
-            snapshot.epoch
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn record_publishes_snapshot_and_hints_in_commit_order() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-
-        let state = NodeState::open(config, None)?;
-        let publisher = state.chain_event_publisher();
-        let epoch = publisher.epoch();
-        let hash_a = Hash256::from_le_bytes(&[0xAA; 32]);
-        let hash_b = Hash256::from_le_bytes(&[0xBB; 32]);
-
-        let first = publisher.record(HintKind::Connected, 1, hash_a);
-        let second = publisher.record(HintKind::Disconnected, 0, hash_b);
-
-        let expected_first = ChainEventHint {
-            kind: HintKind::Connected,
-            height: 1,
-            hash: hash_a,
-            epoch,
-            sequence: 1,
-        };
-        let expected_second = ChainEventHint {
-            kind: HintKind::Disconnected,
-            height: 0,
-            hash: hash_b,
-            epoch,
-            sequence: 2,
-        };
-        assert_eq!(first, expected_first);
-        assert_eq!(second, expected_second);
-        assert_eq!(
-            publisher.snapshot(),
-            state.active_chain_snapshot(),
-            "node and publisher expose one snapshot view"
-        );
-        assert_eq!(
-            state.active_chain_snapshot(),
-            ChainSnapshot {
-                epoch,
-                sequence: 2,
-                tip_hash: hash_b,
-                tip_height: 0,
-            },
-            "the last committed event wins the snapshot cell"
-        );
-
-        let rx = state.chain_event_hints();
-        let mut hints = Vec::new();
-        let hint_rx = rx.lock();
-        while let Ok(hint) = hint_rx.try_recv() {
-            hints.push(hint);
-        }
-        assert_eq!(
-            hints,
-            vec![expected_first, expected_second],
-            "one hint per committed event, in commit order"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn record_drops_hints_when_channel_full() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = dir.path().join("node");
-        config.p2p.listen.clear();
-
-        let state = NodeState::open(config, None)?;
-        let publisher = state.chain_event_publisher();
-        let rx = state.chain_event_hints();
-        for sequence in 1..=super::CHAIN_HINT_CHANNEL_LIMIT {
-            let sequence = u64::try_from(sequence)?;
-            let mut tip = [0_u8; 32];
-            tip[..8].copy_from_slice(&sequence.to_le_bytes());
-            publisher.record(
-                HintKind::Connected,
-                u32::try_from(sequence)?,
-                Hash256::from_le_bytes(&tip),
-            );
-        }
-
-        // The next record finds a full channel: the hint is dropped by
-        // design, while the commit itself still lands and the snapshot
-        // still advances. Dropping never blocks or fails the commit path.
-        let overflow = publisher.record(
-            HintKind::Connected,
-            u32::try_from(super::CHAIN_HINT_CHANNEL_LIMIT)?,
-            Hash256::from_le_bytes(&[0xFF; 32]),
-        );
-        assert_eq!(
-            overflow.sequence,
-            u64::try_from(super::CHAIN_HINT_CHANNEL_LIMIT)? + 1,
-            "the record itself is sequenced even when its hint is dropped"
-        );
-        assert_eq!(
-            state.active_chain_snapshot(),
-            ChainSnapshot {
-                epoch: publisher.epoch(),
-                sequence: overflow.sequence,
-                tip_hash: Hash256::from_le_bytes(&[0xFF; 32]),
-                tip_height: u32::try_from(super::CHAIN_HINT_CHANNEL_LIMIT)?,
-            },
-        );
-
-        let mut drained = Vec::new();
-        let drain_rx = rx.lock();
-        while let Ok(hint) = drain_rx.try_recv() {
-            drained.push(hint.sequence);
-        }
-        assert_eq!(
-            drained.len(),
-            super::CHAIN_HINT_CHANNEL_LIMIT,
-            "exactly the bounded hints are queued"
-        );
-        assert_eq!(drained[0], 1, "the oldest hint survives at the front");
-        assert_eq!(
-            drained.last().copied(),
-            Some(u64::try_from(super::CHAIN_HINT_CHANNEL_LIMIT)?),
-            "the overflow hint is the one that was dropped"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn corrupt_process_epoch_file_refuses_start() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let data_dir = dir.path().join("node");
-        std::fs::create_dir_all(&data_dir)?;
-        std::fs::write(data_dir.join("process-epoch"), b"seven\n")?;
-
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = data_dir;
-        config.p2p.listen.clear();
-
-        let Err(error) = NodeState::open(config, None) else {
-            anyhow::bail!("a corrupt process-epoch file must refuse startup");
-        };
-        assert!(
-            error.to_string().contains("process-epoch"),
-            "the refusal names the corrupt file: {error}"
-        );
-        assert_eq!(
-            std::fs::read(dir.path().join("node").join("process-epoch"))?,
-            b"seven\n",
-            "the refusal must not reset the persisted epoch"
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn process_epoch_allocation_is_unique_across_processes() -> anyhow::Result<()> {
-        const CHILD_DIR_ENV: &str = "BITCOIN_RS_TEST_EPOCH_CHILD_DIR";
-        const CHILDREN: usize = 8;
-        // Subprocess mode: this same test binary was re-exec'd by the parent
-        // below. Park on the start barrier, then allocate one epoch.
-        if let Ok(data_dir) = std::env::var(CHILD_DIR_ENV) {
-            let data_dir = std::path::PathBuf::from(data_dir);
-            let dir = cap_std::fs::Dir::open_ambient_dir(&data_dir, cap_std::ambient_authority())?;
-            std::fs::write(data_dir.join(format!("ready-{}", std::process::id())), b"")?;
-            let go = data_dir.join("go");
-            let deadline = std::time::Instant::now() + std::time::Duration::from_mins(1);
-            while !go.exists() {
-                if std::time::Instant::now() >= deadline {
-                    anyhow::bail!("epoch child never saw the start barrier");
-                }
-                std::thread::sleep(std::time::Duration::from_millis(5));
-            }
-            let epoch = super::allocate_process_epoch(&dir)?;
-            std::fs::write(
-                data_dir.join(format!("epoch-{}", std::process::id())),
-                format!("{epoch}\n"),
-            )?;
-            std::process::exit(0);
-        }
-
-        let dir = tempfile::tempdir()?;
-        let data_dir = dir.path().join("node");
-        std::fs::create_dir_all(&data_dir)?;
-
-        // Harness test names are crate-relative; `module_path!()` is not.
-        let test_name = concat!(
-            module_path!(),
-            "::process_epoch_allocation_is_unique_across_processes"
-        )
-        .split("::")
-        .skip(1)
-        .collect::<Vec<_>>()
-        .join("::");
-        let exe = std::env::current_exe()?;
-        let mut children = Vec::new();
-        for _ in 0..CHILDREN {
-            children.push(
-                std::process::Command::new(&exe)
-                    .args(["--exact", &test_name])
-                    .env(CHILD_DIR_ENV, &data_dir)
-                    .stdout(std::process::Stdio::piped())
-                    .stderr(std::process::Stdio::piped())
-                    .spawn()?,
-            );
-        }
-
-        // Every child must be alive and parked before any allocates, so all
-        // eight genuinely contend for the same lock file on one data dir.
-        let deadline = std::time::Instant::now() + std::time::Duration::from_mins(1);
-        loop {
-            let ready = std::fs::read_dir(&data_dir)?
-                .filter_map(Result::ok)
-                .filter(|entry| entry.file_name().to_string_lossy().starts_with("ready-"))
-                .count();
-            if ready == CHILDREN {
-                break;
-            }
-            if std::time::Instant::now() >= deadline {
-                anyhow::bail!("epoch children never became ready: {ready}/{CHILDREN}");
-            }
-            std::thread::sleep(std::time::Duration::from_millis(5));
-        }
-        std::fs::write(data_dir.join("go"), b"")?;
-
-        let mut epochs = Vec::new();
-        for child in children {
-            let output = child.wait_with_output()?;
-            anyhow::ensure!(
-                output.status.success(),
-                "epoch child failed: {}",
-                String::from_utf8_lossy(&output.stderr)
-            );
-        }
-        for entry in std::fs::read_dir(&data_dir)? {
-            let entry = entry?;
-            if entry.file_name().to_string_lossy().starts_with("epoch-") {
-                let text = std::fs::read_to_string(entry.path())?;
-                epochs.push(text.trim().parse::<u64>()?);
-            }
-        }
-        epochs.sort_unstable();
-        assert_eq!(
-            epochs.len(),
-            CHILDREN,
-            "each child reports exactly one epoch"
-        );
-        for (index, epoch) in epochs.iter().enumerate() {
-            assert_eq!(
-                *epoch,
-                u64::try_from(index)? + 1,
-                "concurrent children must each own one distinct epoch: {epochs:?}"
-            );
-        }
-        assert_eq!(
-            std::fs::read_to_string(data_dir.join("process-epoch"))?,
-            format!("{}\n", epochs[CHILDREN - 1]),
-            "the persisted file must name the highest allocated epoch"
-        );
-        Ok(())
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn symlinked_epoch_lock_refuses_start() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let data_dir = dir.path().join("node");
-        std::fs::create_dir_all(&data_dir)?;
-        std::fs::write(data_dir.join("process-epoch"), b"41\n")?;
-        std::os::unix::fs::symlink("process-epoch", data_dir.join(".process-epoch.lock"))?;
-
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = data_dir.clone();
-        config.p2p.listen.clear();
-
-        let Err(error) = NodeState::open(config, None) else {
-            anyhow::bail!("a symlinked epoch lock must refuse startup");
-        };
-        assert!(
-            error.to_string().contains("process epoch lock"),
-            "the refusal names the lock target: {error:#}"
-        );
-        assert_eq!(
-            std::fs::read_to_string(data_dir.join("process-epoch"))?,
-            "41\n",
-            "the symlink must not be followed and the epoch must not reset"
-        );
-        Ok(())
-    }
-
-    #[cfg(all(unix, not(target_vendor = "apple")))]
-    #[test]
-    #[cfg(not(target_vendor = "apple"))]
-    fn non_regular_epoch_lock_refuses_start() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let data_dir = dir.path().join("node");
-        std::fs::create_dir_all(&data_dir)?;
-        std::fs::write(data_dir.join("process-epoch"), b"7\n")?;
-        let lock_dir = cap_std::fs::Dir::open_ambient_dir(&data_dir, cap_std::ambient_authority())?;
-        rustix::fs::mkfifoat(
-            &lock_dir,
-            ".process-epoch.lock",
-            rustix::fs::Mode::from_raw_mode(0o600),
-        )?;
-
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = data_dir.clone();
-        config.p2p.listen.clear();
-
-        let Err(error) = NodeState::open(config, None) else {
-            anyhow::bail!("a non-regular epoch lock must refuse startup");
-        };
-        assert!(
-            error.to_string().contains("not a regular file"),
-            "the refusal names the lock type: {error:#}"
-        );
-        assert_eq!(
-            std::fs::read_to_string(data_dir.join("process-epoch"))?,
-            "7\n",
-            "the refusal must not reset the persisted epoch"
-        );
-        Ok(())
-    }
-
-    fn mined_regtest_child(prev_blockhash: BlockHash) -> anyhow::Result<Block> {
-        mined_regtest_child_at(prev_blockhash, 1_296_688_603, 1)
-    }
-
-    fn mined_regtest_child_at(
-        prev_blockhash: BlockHash,
-        time: u32,
-        height: u32,
-    ) -> anyhow::Result<Block> {
-        let mut script_sig = vec![8];
-        script_sig.extend_from_slice(&height.to_le_bytes());
-        script_sig.extend_from_slice(&time.to_le_bytes());
-        let coinbase = Tx {
-            version: 2,
-            lock_time: 0,
-            inputs: vec![TxIn {
-                previous_output: OutPoint::new(Txid::default(), u32::MAX),
-                script_sig,
-                sequence: u32::MAX,
-                witness: Vec::new(),
-            }],
-            outputs: vec![TxOut {
-                value: 1,
-                script_pubkey: Vec::new(),
-            }],
-        };
-        let mut block = Block {
-            header: Header {
-                version: 1,
-                prev_blockhash,
-                merkle_root: Hash256::default(),
-                time,
-                bits: 0x207f_ffff,
-                nonce: 0,
-            },
-            txs: vec![coinbase],
-        };
-        block.header.merkle_root = merkle_root(&block.txs)
-            .ok_or_else(|| std::io::Error::other("test block has no merkle root"))?;
-        while !pow_met(block.header.bits, block.block_hash().0) {
-            block.header.nonce = block
-                .header
-                .nonce
-                .checked_add(1)
-                .ok_or_else(|| std::io::Error::other("test nonce exhausted"))?;
-        }
-        Ok(block)
-    }
-
-    /// Pairwise double-SHA256 fold over little-endian txid bytes, duplicating
-    /// the last leaf on odd levels (the native stand-in for
-    /// `compute_merkle_root`).
-    fn merkle_root(txs: &[Tx]) -> Option<Hash256> {
-        let mut leaves: Vec<[u8; 32]> = txs.iter().map(|tx| *tx.txid().as_bytes()).collect();
-        if leaves.is_empty() {
-            return None;
-        }
-        while leaves.len() > 1 {
-            let original_len = leaves.len();
-            let mut next = Vec::with_capacity(original_len.div_ceil(2));
-            for pos in 0..original_len.div_ceil(2) {
-                let left = leaves[2 * pos];
-                let right = leaves[(2 * pos + 1).min(original_len - 1)];
-                let mut pair = [0_u8; 64];
-                pair[..32].copy_from_slice(&left);
-                pair[32..].copy_from_slice(&right);
-                next.push(double_sha256(&pair).to_le_bytes());
-            }
-            leaves = next;
-        }
-        Some(Hash256::from_le_bytes(&leaves[0]))
-    }
-
-    /// Regtest-easy compact-target `PoW` check over the hash as a 256-bit
-    /// little-endian integer (mirrors `chain::pow::compact_is_met_by` for the
-    /// >3-exponent, 3-byte-mantissa forms these fixtures mine).
-    fn pow_met(bits: u32, hash: Hash256) -> bool {
-        let exponent = u8::try_from(bits >> 24).unwrap_or(0);
-        let mantissa = bits & 0x007f_ffff;
-        if exponent <= 3 || exponent > 32 || mantissa > 0x00ff_ffff {
-            return false;
-        }
-        let bytes = hash.as_byte_array();
-        let low = usize::from(exponent - 3);
-        let window = u32::from(bytes[low])
-            | u32::from(bytes[low + 1]) << 8
-            | u32::from(bytes[low + 2]) << 16;
-        window <= mantissa && bytes[usize::from(exponent)..].iter().all(|&byte| byte == 0)
-    }
-
-    // -----------------------------------------------------------------------
-    // A2 cycle 3: witness is published only after CURRENT root fsync
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn witness_is_published_only_after_current_root_sync() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let data_dir = dir.path().join("node");
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = data_dir.clone();
-        config.p2p.listen.clear();
-
-        let state = NodeState::open(config.clone(), None)?;
-        let genesis = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        let tip = state.apply_block(&genesis)?;
-
-        // Publish a checkpoint — witness must be written after Published.
-        state.publish_checkpoint()?;
-        let witness_path = data_dir.join("applied-tip-witness.json");
-        assert!(
-            witness_path.exists(),
-            "witness file must exist after checkpoint publication"
-        );
-        let genesis_hex = config.network.genesis_block_hash().to_string_be();
-        let witness = crate::recovery_evidence::read_witness(&data_dir, &genesis_hex)
-            .ok_or_else(|| anyhow::anyhow!("witness must be readable"))?;
-        assert_eq!(witness.height, tip.height);
-        assert_eq!(witness.block_hash, tip.hash.to_string_be());
-        drop(state);
-
-        // Now inject a failpoint at CurrentRootSync — the last stage before
-        // the checkpoint is considered Published. The checkpoint must fail,
-        // and no new witness must be written for the failed checkpoint.
-        let dir2 = tempfile::tempdir()?;
-        let data_dir2 = dir2.path().join("node");
-        let mut config2 = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config2.data_dir = data_dir2.clone();
-        config2.p2p.listen.clear();
-
-        let state2 = NodeState::open(config2.clone(), None)?;
-        let genesis2 = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        state2.apply_block(&genesis2)?;
-
-        // Inject failpoint at the final root fsync — checkpoint fails before
-        // returning Published, so no witness should be written.
-        crate::checkpoint::inject_next_checkpoint_failpoint(
-            crate::checkpoint::CheckpointFailpoint::CurrentRootSync,
-        );
-        let result = state2.publish_checkpoint();
-        assert!(
-            result.is_err(),
-            "checkpoint must fail when CurrentRootSync fails"
-        );
-        let witness_path2 = data_dir2.join("applied-tip-witness.json");
-        assert!(
-            !witness_path2.exists(),
-            "no witness must be written when checkpoint fails before publication"
-        );
-        Ok(())
-    }
-
-    // -----------------------------------------------------------------------
-    // #208: a checkpoint restore far behind the durable witness must be loud
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn stale_checkpoint_restore_surfaces_warning_not_silence() -> anyhow::Result<()> {
-        let dir = tempfile::tempdir()?;
-        let data_dir = dir.path().join("node");
-        let mut config = crate::NodeConfig::default_for_network(crate::Network::Regtest);
-        config.data_dir = data_dir.clone();
-        config.p2p.listen.clear();
-        config.indexes.script_index = crate::config::ScriptIndexMode::Disabled;
-
-        // Apply genesis and publish a checkpoint at height 0.
-        let state = NodeState::open(config.clone(), None)?;
-        let genesis = bitcoin_rs_primitives::Network::Regtest.genesis_block();
-        state.apply_block(&genesis)?;
-        state.write_clean_checkpoint()?;
-        drop(state);
-
-        // Simulate the #208 scenario: the node previously ran far ahead
-        // (height 5000) and published a checkpoint there, writing a witness
-        // at that height. A crash or clean stop left the checkpoint tree
-        // pinned at height 0 while the witness records height 5000.
-        let genesis_hex = config.network.genesis_block_hash().to_string_be();
-        let stale_witness = crate::recovery_evidence::AppliedTipWitness::new(
-            genesis_hex,
-            1, // older epoch
-            5000,
-            "cccc",
-            1000,
-        );
-        crate::recovery_evidence::write_witness(&data_dir, &stale_witness)?;
-
-        // Reopen: the checkpoint at height 0 is restored, the witness at
-        // 5000 triggers checkpoint-fallback detection. The warning store
-        // must carry the fallback warning — the restore must not be silent.
-        let resumed = NodeState::open(config.clone(), None)?;
-        let warnings = resumed.warning_store().warnings();
-        assert!(
-            !warnings.is_empty(),
-            "a stale checkpoint restore 5000 blocks behind the witness must \
-             produce at least one rollback warning, not silence"
-        );
-        assert!(
-            warnings.iter().any(|w| w.contains("height 5000")),
-            "the warning must name the witness height; got: {warnings:?}"
-        );
-        assert_eq!(
-            resumed.resume_source(),
-            ResumeSource::Checkpoint,
-            "the checkpoint is still accepted — it is valid, just stale"
-        );
-        Ok(())
-    }
-}
+mod tests;
